==> 5_table_creation.R <==
####### GETTING DATA TABLES FROM RESULTS ####################################

#### loading packages
library(tidyr)
library(data.table)
library(dplyr)
library(stringr)
library(ggplot2)

##### read in data

#### !!! update "XXXXX" to e.g. "C:/Users..." depending on where project saved
myfiles = list.files(path="C:/Users/nichola.naylor/Documents/WHO_2023/VAF_AMR_EconBurden/outputs/fulloutput_chunks/",
                     pattern="*.RData", full.names = TRUE)

### !!! in future iterations would be good to explore why NA scenarios come up
### where there are unit costs sampled + no cases - remove earlier on so don't have to
### keep removing in table/plot creation

### split into the ones that need adjusting and those that don't
### compared to WHO tables - don't need to do pseudo or strep syndrome specifications
### as using all modelled syndromes
### do need to add hib and strep scenarios that = 1 scenario 

#### there is probaby a more efficient way of doing but trying to adapt into older code for now
#### note that the numbers used in the functions would need updating to adapt for different scenarios
#### could try to get 1 function where you just specify the group_by variables then just run that
### for each

adj <- c("C:/Users/nichola.naylor/Documents/WHO_2023/VAF_AMR_EconBurden/outputs/fulloutput_chunks/Stre31.RData",
         "C:/Users/nichola.naylor/Documents/WHO_2023/VAF_AMR_EconBurden/outputs/fulloutput_chunks/Stre32.RData",
         "C:/Users/nichola.naylor/Documents/WHO_2023/VAF_AMR_EconBurden/outputs/fulloutput_chunks/Stre33.RData",
         "C:/Users/nichola.naylor/Documents/WHO_2023/VAF_AMR_EconBurden/outputs/fulloutput_chunks/Stre34.RData",
         "C:/Users/nichola.naylor/Documents/WHO_2023/VAF_AMR_EconBurden/outputs/fulloutput_chunks/Stre35.RData",
         "C:/Users/nichola.naylor/Documents/WHO_2023/VAF_AMR_EconBurden/outputs/fulloutput_chunks/Stre36.RData",
         "C:/Users/nichola.naylor/Documents/WHO_2023/VAF_AMR_EconBurden/outputs/fulloutput_chunks/Stre37.RData",
         "C:/Users/nichola.naylor/Documents/WHO_2023/VAF_AMR_EconBurden/outputs/fulloutput_chunks/Stre38.RData",
       "C:/Users/nicho/OneDrive/Documents/VAF_AMR_EconBurden/outputs/fulloutput_chunks/resultsHaem15.RData",
       "C:/Users/nicho/OneDrive/Documents/VAF_AMR_EconBurden/outputs/fulloutput_chunks/resultsHaem16.RData")

nonadj <- setdiff(myfiles,adj) ## find ones that don't need adjusting
       

#### MEDIAN & IQR INSTEAD #################

mediqr1 <- function(vaccine_output_dt){
  vaccine_output_dt[WHO.Region=="PAHO",WHO.Region := "AMRO"]
  x <- vaccine_output_dt%>% group_by(WHO.Region, Pathogen, Infectious.syndrome,
                                                vaccine_id, run) %>%
    summarise(who_region_cost_A = sum(avertable_cost_cases, na.rm=TRUE),
              who_region_cost_T = sum(cost_cases, na.rm=TRUE),
              who_region_days_A = sum(avertable_days_cases, na.rm=TRUE),
              who_region_days_T = sum(days_cases, na.rm=TRUE))  %>%
  group_by(WHO.Region, Pathogen, Infectious.syndrome, vaccine_id) %>%
    summarise(median_avert_costing = median(who_region_cost_A, na.rm=TRUE),
              LOWIQR_avert_costing = quantile(who_region_cost_A,0.25, na.rm = TRUE),
              HIGHIQR_avert_costing = quantile(who_region_cost_A,0.75, na.rm = TRUE),
              median_total_costing = median(who_region_cost_T, na.rm=TRUE),
              LOWIQR_total_costing = quantile(who_region_cost_T,0.25, na.rm = TRUE),
              HIGHIQR_total_costing = quantile(who_region_cost_T,0.75, na.rm = TRUE),
              median_avert_days = median(who_region_days_A, na.rm=TRUE),
              LOWIQR_avert_days = quantile(who_region_days_A,0.25, na.rm = TRUE),
              HIGHIQR_avert_days = quantile(who_region_days_A,0.75, na.rm = TRUE),
              median_total_days = median(who_region_days_T, na.rm=TRUE),
              LOWIQR_total_days = quantile(who_region_days_T,0.25, na.rm = TRUE),
              HIGHIQR_total_days = quantile(who_region_days_T,0.75, na.rm = TRUE)) %>%
    mutate(across(median_avert_costing:HIGHIQR_total_days,
                  ~ format(., big.mark = ",", scientific = F)))%>%
    mutate( median_iqr_total_cost = str_c(median_total_costing, " ( ", LOWIQR_total_costing, "-",
                                       HIGHIQR_total_costing," ) "),
            median_iqr_averted_cost = str_c(median_avert_costing, " ( ", LOWIQR_avert_costing, "-",
                                         HIGHIQR_avert_costing," ) "),
            median_iqr_total_days = str_c(median_total_days, " ( ", LOWIQR_total_days, "-",
                                       HIGHIQR_total_days," ) "),
            median_iqr_averted_days = str_c(median_avert_days, " ( ", LOWIQR_avert_days, "-",
                                         HIGHIQR_avert_days," ) ")) %>%
    select(!c("median_avert_costing",  "LOWIQR_avert_costing",
                                 "HIGHIQR_avert_costing", "median_total_costing" ,
                                 "LOWIQR_total_costing" , "HIGHIQR_total_costing" ,
                                 "median_avert_days",  "LOWIQR_avert_days",
                                 "HIGHIQR_avert_days", "median_total_days" ,
                                 "LOWIQR_total_days" , "HIGHIQR_total_days")) %>%
    
  as.data.table()
  return(x)
}
########### nonadj ################
hospital_cost_l <- list()
for (i in 1:length(nonadj)){
  load(nonadj[i])
  vaccine_output_dt[WHO.Region=="PAHO", WHO.Region := "AMRO"]
  hospital_cost_l[[i]] <- mediqr1(vaccine_output_dt)
}

########### adj ################
hospital_cost_l_adj <- list()

## note that strep and hib results that are Current coverage are not included in below 

##### loading the datasets we need
### adding dataset ref by hand re. numbers so would need checking if updated 
### scenarios !!!
load("outputs/fulloutput_chunks/Stre37.RData") 
serotype.58 <- vaccine_output_dt
load("outputs/fulloutput_chunks/Stre35.RData") 
serotype <- rbind(vaccine_output_dt,serotype.58)
rm(serotype.58)
## ?? coming up with error but still creates the right data table ??
## updating vaccine id to match to enable use of function
unique(serotype$vaccine_id)
serotype[ , vaccine_id := "Streptococcus pneumoniae_both_0.9_5 years_BSI, CNS infections, Cardiac infections, LRI_6, 10, 14 weeks"]
hospital_cost_l_adj[[1]] <- mediqr1(serotype)


load("outputs/fulloutput_chunks/Stre38.RData") 
serotype.58.eld <- vaccine_output_dt
load("outputs/fulloutput_chunks/Stre36.RData") 
serotype.eld <- rbind(vaccine_output_dt, serotype.58.eld)
rm(serotype.58.eld)
unique(serotype.eld$vaccine_id)
serotype.eld <- serotype.eld[!is.na(vaccine_id)] ## remove those not associated with any vaccine scenarios
serotype.eld[ , vaccine_id := "Streptococcus pneumoniae_both_0.9_5 years_BSI, CNS infections, Cardiac infections, LRI_6, 10, 14 weeks & elderly age group"]
hospital_cost_l_adj[[2]] <- mediqr1(serotype.eld)

load("outputs/fulloutput_chunks/Stre33.RData") 
improv.7 <- vaccine_output_dt
load("outputs/fulloutput_chunks/Stre31.RData")
improv <- rbind(vaccine_output_dt, improv.7)
rm(improv.7)
unique(improv$vaccine_id)
improv <- improv[!is.na(vaccine_id)] ## remove those not associated with any vaccine scenarios
improv[ , vaccine_id := "Streptococcus pneumoniae - Improved_both_0.9_5 years_BSI, CNS infections, Cardiac infections, LRI_6 weeks"]
hospital_cost_l_adj[[3]] <- mediqr1(improv)

load("outputs/fulloutput_chunks/Stre32.RData") 
improv.5.eld <- vaccine_output_dt
load("outputs/fulloutput_chunks/Stre34.RData") 
improv.eld <- rbind(vaccine_output_dt,improv.5.eld)
rm(improv.5.eld)
unique(improv.eld$vaccine_id)
improv.eld <- improv.eld[!is.na(vaccine_id)] ## remove those not associated with any vaccine scenarios
improv.eld[ , vaccine_id := "Streptococcus pneumoniae - Improved_both_0.9_5 years_BSI, CNS infections, Cardiac infections, LRI_6 weeks & elderly age group"]
hospital_cost_l_adj[[4]] <- mediqr1(improv.eld)


### HiB
load("outputs/fulloutput_chunks/Haem16.RData") 
Hib_93 <- vaccine_output_dt
load("outputs/fulloutput_chunks/Haem15.RData")
Hib <- rbind(vaccine_output_dt,Hib_93)
rm(Hib_93)
unique(Hib$vaccine_id)
Hib <- Hib[!is.na(vaccine_id)] ## remove those not associated with any vaccine scenarios
Hib[ , vaccine_id := "Haemophilus influenzae type B_both_0.9_5 years_All_6, 10, 14 weeks"]
hospital_cost_l_adj[[5]] <- mediqr1(Hib)


hospital_c <- rbindlist(hospital_cost_l)
hospital_cA <- rbindlist(hospital_cost_l_adj)
hospital_c <- rbind(hospital_c,hospital_cA)

###!!! not including those to mapped to any vaccine ID
hospital_c <- hospital_c[vaccine_id!="_NA_NA___"]
hospital_c <- hospital_c[Pathogen!="NA"]

hospital_c <- dcast(hospital_c, Pathogen + Infectious.syndrome +
                      vaccine_id ~ WHO.Region, value.var = c( "median_iqr_total_cost" 
                                                              ,"median_iqr_averted_cost"
                                                              ,"median_iqr_total_days"  
                                                              ,"median_iqr_averted_days"))

hospital_c[, c("vaccine_target_disease","efficacy" ,"coverage"  ,
               "duration", "target disease","target population") := tstrsplit(vaccine_id, "_", fixed=TRUE)]



write.csv(hospital_c, file="outputs/END_hospital_costs_output_median.csv")


######## REGION + ALL LEVELS ############
mediqr1.all <- function(vaccine_output_dt){
  x <- vaccine_output_dt%>% group_by(WHO.Region, Pathogen, class, Infectious.syndrome,
                                     vaccine_id, run)  %>%
    summarise(who_region_cost_A = sum(avertable_cost_cases, na.rm=TRUE),
              who_region_cost_T = sum(cost_cases, na.rm=TRUE),
              who_region_days_A = sum(avertable_days_cases, na.rm=TRUE),
              who_region_days_T = sum(days_cases, na.rm=TRUE))  %>%
    group_by(WHO.Region, Pathogen, class, Infectious.syndrome,
             vaccine_id)  %>%
    summarise(median_avert_costing = median(who_region_cost_A, na.rm=TRUE),
              LOWIQR_avert_costing = quantile(who_region_cost_A,0.25, na.rm = TRUE),
              HIGHIQR_avert_costing = quantile(who_region_cost_A,0.75, na.rm = TRUE),
              median_total_costing = median(who_region_cost_T, na.rm=TRUE),
              LOWIQR_total_costing = quantile(who_region_cost_T,0.25, na.rm = TRUE),
              HIGHIQR_total_costing = quantile(who_region_cost_T,0.75, na.rm = TRUE),
              median_avert_days = median(who_region_days_A, na.rm=TRUE),
              LOWIQR_avert_days = quantile(who_region_days_A,0.25, na.rm = TRUE),
              HIGHIQR_avert_days = quantile(who_region_days_A,0.75, na.rm = TRUE),
              median_total_days = median(who_region_days_T, na.rm=TRUE),
              LOWIQR_total_days = quantile(who_region_days_T,0.25, na.rm = TRUE),
              HIGHIQR_total_days = quantile(who_region_days_T,0.75, na.rm = TRUE)) %>%
    mutate(across(median_avert_costing:HIGHIQR_total_days,
                  ~ format(., big.mark = ",", scientific = F)))%>%
    mutate( median_iqr_total_cost = str_c(median_total_costing, " ( ", LOWIQR_total_costing, "-",
                                          HIGHIQR_total_costing," ) "),
            median_iqr_averted_cost = str_c(median_avert_costing, " ( ", LOWIQR_avert_costing, "-",
                                            HIGHIQR_avert_costing," ) "),
            median_iqr_total_days = str_c(median_total_days, " ( ", LOWIQR_total_days, "-",
                                          HIGHIQR_total_days," ) "),
            median_iqr_averted_days = str_c(median_avert_days, " ( ", LOWIQR_avert_days, "-",
                                            HIGHIQR_avert_days," ) ")) %>%
    select(!c("median_avert_costing",  "LOWIQR_avert_costing",
              "HIGHIQR_avert_costing", "median_total_costing" ,
              "LOWIQR_total_costing" , "HIGHIQR_total_costing" ,
              "median_avert_days",  "LOWIQR_avert_days",
              "HIGHIQR_avert_days", "median_total_days" ,
              "LOWIQR_total_days" , "HIGHIQR_total_days")) %>%
    
    as.data.table()
  return(x)
}
########### nonadj ################
hospital_cost_l <- list()
for (i in 1:length(nonadj)){
  load(nonadj[i])
  vaccine_output_dt[WHO.Region=="PAHO", WHO.Region := "AMRO"]
  hospital_cost_l[[i]] <- mediqr1.all(vaccine_output_dt)
}

########### adj ################
hospital_cost_l_adj <- list()

hospital_cost_l_adj[[1]] <- mediqr1.all(serotype)
hospital_cost_l_adj[[2]] <- mediqr1.all(serotype.eld)
hospital_cost_l_adj[[3]] <- mediqr1.all(improv)
hospital_cost_l_adj[[4]] <- mediqr1.all(improv.eld)
hospital_cost_l_adj[[5]] <- mediqr1.all(Hib)


hospital_c <- rbindlist(hospital_cost_l)
hospital_cA <- rbindlist(hospital_cost_l_adj)
hospital_c <- rbind(hospital_c,hospital_cA)
hospital_c[, c("vaccine_target_disease","efficacy" ,"coverage"  ,
               "duration", "target disease","target population") := tstrsplit(vaccine_id, "_", fixed=TRUE)]

###!!! not including those to mapped to any vaccine ID
hospital_c <- hospital_c[vaccine_id!="_NA_NA___"]
hospital_c <- hospital_c[Pathogen!="NA"]

save(hospital_c,file="outputs/hospitalC_region_averted.RData")


######## REGION + CLASS ############

mediqr.reg.class <- function(vaccine_output_dt){
  x <- vaccine_output_dt%>% group_by(WHO.Region, class,
                                     vaccine_id, run) %>%
    summarise(who_region_cost_A = sum(avertable_cost_cases, na.rm=TRUE),
              who_region_cost_T = sum(cost_cases, na.rm=TRUE),
              who_region_days_A = sum(avertable_days_cases, na.rm=TRUE),
              who_region_days_T = sum(days_cases, na.rm=TRUE))  %>%
    group_by(WHO.Region, class,
             vaccine_id) %>%
    summarise(median_avert_costing = median(who_region_cost_A, na.rm=TRUE),
              LOWIQR_avert_costing = quantile(who_region_cost_A,0.25, na.rm = TRUE),
              HIGHIQR_avert_costing = quantile(who_region_cost_A,0.75, na.rm = TRUE),
              median_total_costing = median(who_region_cost_T, na.rm=TRUE),
              LOWIQR_total_costing = quantile(who_region_cost_T,0.25, na.rm = TRUE),
              HIGHIQR_total_costing = quantile(who_region_cost_T,0.75, na.rm = TRUE),
              median_avert_days = median(who_region_days_A, na.rm=TRUE),
              LOWIQR_avert_days = quantile(who_region_days_A,0.25, na.rm = TRUE),
              HIGHIQR_avert_days = quantile(who_region_days_A,0.75, na.rm = TRUE),
              median_total_days = median(who_region_days_T, na.rm=TRUE),
              LOWIQR_total_days = quantile(who_region_days_T,0.25, na.rm = TRUE),
              HIGHIQR_total_days = quantile(who_region_days_T,0.75, na.rm = TRUE)) %>%
    mutate(across(median_avert_costing:HIGHIQR_total_days,
                  ~ format(., big.mark = ",", scientific = F)))%>%
    mutate( median_iqr_total_cost = str_c(median_total_costing, " ( ", LOWIQR_total_costing, "-",
                                          HIGHIQR_total_costing," ) "),
            median_iqr_averted_cost = str_c(median_avert_costing, " ( ", LOWIQR_avert_costing, "-",
                                            HIGHIQR_avert_costing," ) "),
            median_iqr_total_days = str_c(median_total_days, " ( ", LOWIQR_total_days, "-",
                                          HIGHIQR_total_days," ) "),
            median_iqr_averted_days = str_c(median_avert_days, " ( ", LOWIQR_avert_days, "-",
                                            HIGHIQR_avert_days," ) ")) %>%
    select(!c("median_avert_costing",  "LOWIQR_avert_costing",
              "HIGHIQR_avert_costing", "median_total_costing" ,
              "LOWIQR_total_costing" , "HIGHIQR_total_costing" ,
              "median_avert_days",  "LOWIQR_avert_days",
              "HIGHIQR_avert_days", "median_total_days" ,
              "LOWIQR_total_days" , "HIGHIQR_total_days")) %>%
    
    as.data.table()
  return(x)
}
########### nonadj ################
hospital_cost_l <- list()
for (i in 1:length(nonadj)){
  load(nonadj[i])
  vaccine_output_dt[WHO.Region=="PAHO", WHO.Region := "AMRO"]
  hospital_cost_l[[i]] <- mediqr.reg.class(vaccine_output_dt)
}

########### adj ################
hospital_cost_l_adj <- list()

hospital_cost_l_adj[[1]] <- mediqr.reg.class(serotype)
hospital_cost_l_adj[[2]] <- mediqr.reg.class(serotype.eld)
hospital_cost_l_adj[[3]] <- mediqr.reg.class(improv)
hospital_cost_l_adj[[4]] <- mediqr.reg.class(improv.eld)
hospital_cost_l_adj[[5]] <- mediqr.reg.class(Hib)


hospital_c <- rbindlist(hospital_cost_l)
hospital_cA <- rbindlist(hospital_cost_l_adj)
hospital_c <- rbind(hospital_c,hospital_cA)

hospital_c[, c("vaccine_target_disease","efficacy" ,"coverage"  ,
               "duration", "target disease","target population") := tstrsplit(vaccine_id, "_", fixed=TRUE)]

###!!! not including those to mapped to any vaccine ID
hospital_c <- hospital_c[vaccine_id!="_NA_NA___"]

save(hospital_c,file="outputs/hospitalC_region_averted_class.RData")


######## REGION + PATHOGEN ############

mediqr.reg.path <- function(vaccine_output_dt){
  x <- vaccine_output_dt%>% group_by(WHO.Region, Pathogen, 
                             vaccine_id, run) %>%
    summarise(who_region_cost_A = sum(avertable_cost_cases, na.rm=TRUE),
              who_region_cost_T = sum(cost_cases, na.rm=TRUE),
              who_region_days_A = sum(avertable_days_cases, na.rm=TRUE),
              who_region_days_T = sum(days_cases, na.rm=TRUE))  %>%
    group_by(WHO.Region, Pathogen, 
             vaccine_id) %>%
    summarise(median_avert_costing = median(who_region_cost_A, na.rm=TRUE),
              LOWIQR_avert_costing = quantile(who_region_cost_A,0.25, na.rm = TRUE),
              HIGHIQR_avert_costing = quantile(who_region_cost_A,0.75, na.rm = TRUE),
              median_total_costing = median(who_region_cost_T, na.rm=TRUE),
              LOWIQR_total_costing = quantile(who_region_cost_T,0.25, na.rm = TRUE),
              HIGHIQR_total_costing = quantile(who_region_cost_T,0.75, na.rm = TRUE),
              median_avert_days = median(who_region_days_A, na.rm=TRUE),
              LOWIQR_avert_days = quantile(who_region_days_A,0.25, na.rm = TRUE),
              HIGHIQR_avert_days = quantile(who_region_days_A,0.75, na.rm = TRUE),
              median_total_days = median(who_region_days_T, na.rm=TRUE),
              LOWIQR_total_days = quantile(who_region_days_T,0.25, na.rm = TRUE),
              HIGHIQR_total_days = quantile(who_region_days_T,0.75, na.rm = TRUE)) %>%
    mutate(across(median_avert_costing:HIGHIQR_total_days,
                  ~ format(., big.mark = ",", scientific = F)))%>%
    mutate( median_iqr_total_cost = str_c(median_total_costing, " ( ", LOWIQR_total_costing, "-",
                                          HIGHIQR_total_costing," ) "),
            median_iqr_averted_cost = str_c(median_avert_costing, " ( ", LOWIQR_avert_costing, "-",
                                            HIGHIQR_avert_costing," ) "),
            median_iqr_total_days = str_c(median_total_days, " ( ", LOWIQR_total_days, "-",
                                          HIGHIQR_total_days," ) "),
            median_iqr_averted_days = str_c(median_avert_days, " ( ", LOWIQR_avert_days, "-",
                                            HIGHIQR_avert_days," ) ")) %>%
    select(!c("median_avert_costing",  "LOWIQR_avert_costing",
              "HIGHIQR_avert_costing", "median_total_costing" ,
              "LOWIQR_total_costing" , "HIGHIQR_total_costing" ,
              "median_avert_days",  "LOWIQR_avert_days",
              "HIGHIQR_avert_days", "median_total_days" ,
              "LOWIQR_total_days" , "HIGHIQR_total_days")) %>%
    
    as.data.table()
  return(x)
}
########### nonadj ################
hospital_cost_l <- list()
for (i in 1:length(nonadj)){
  load(nonadj[i])
  vaccine_output_dt[WHO.Region=="PAHO", WHO.Region := "AMRO"]
  hospital_cost_l[[i]] <- mediqr.reg.path(vaccine_output_dt)
}

########### adj ################
hospital_cost_l_adj <- list()

hospital_cost_l_adj[[1]] <- mediqr.reg.path(serotype)
hospital_cost_l_adj[[2]] <- mediqr.reg.path(serotype.eld)
hospital_cost_l_adj[[3]] <- mediqr.reg.path(improv)
hospital_cost_l_adj[[4]] <- mediqr.reg.path(improv.eld)
hospital_cost_l_adj[[5]] <- mediqr.reg.path(Hib)


hospital_c <- rbindlist(hospital_cost_l)
hospital_cA <- rbindlist(hospital_cost_l_adj)
hospital_c <- rbind(hospital_c,hospital_cA)

hospital_c[, c("vaccine_target_disease","efficacy" ,"coverage"  ,
               "duration", "target disease","target population") := tstrsplit(vaccine_id, "_", fixed=TRUE)]

###!!! not including those to mapped to any vaccine ID
hospital_c <- hospital_c[vaccine_id!="_NA_NA___"]
hospital_c <- hospital_c[Pathogen!="NA"]

save(hospital_c,file="outputs/hospitalC_region_averted_pathogen.RData")


######## REGION + SYNDROME ############
mediqr.reg.syn <- function(vaccine_output_dt){
  x <- vaccine_output_dt%>% group_by(WHO.Region, Infectious.syndrome,
                                     vaccine_id, run) %>%
    summarise(who_region_cost_A = sum(avertable_cost_cases, na.rm=TRUE),
              who_region_cost_T = sum(cost_cases, na.rm=TRUE),
              who_region_days_A = sum(avertable_days_cases, na.rm=TRUE),
              who_region_days_T = sum(days_cases, na.rm=TRUE))  %>%
    group_by(WHO.Region, Infectious.syndrome,
             vaccine_id) %>%
    summarise(median_avert_costing = median(who_region_cost_A, na.rm=TRUE),
              LOWIQR_avert_costing = quantile(who_region_cost_A,0.25, na.rm = TRUE),
              HIGHIQR_avert_costing = quantile(who_region_cost_A,0.75, na.rm = TRUE),
              median_total_costing = median(who_region_cost_T, na.rm=TRUE),
              LOWIQR_total_costing = quantile(who_region_cost_T,0.25, na.rm = TRUE),
              HIGHIQR_total_costing = quantile(who_region_cost_T,0.75, na.rm = TRUE),
              median_avert_days = median(who_region_days_A, na.rm=TRUE),
              LOWIQR_avert_days = quantile(who_region_days_A,0.25, na.rm = TRUE),
              HIGHIQR_avert_days = quantile(who_region_days_A,0.75, na.rm = TRUE),
              median_total_days = median(who_region_days_T, na.rm=TRUE),
              LOWIQR_total_days = quantile(who_region_days_T,0.25, na.rm = TRUE),
              HIGHIQR_total_days = quantile(who_region_days_T,0.75, na.rm = TRUE)) %>%
    mutate(across(median_avert_costing:HIGHIQR_total_days,
                  ~ format(., big.mark = ",", scientific = F)))%>%
    mutate( median_iqr_total_cost = str_c(median_total_costing, " ( ", LOWIQR_total_costing, "-",
                                          HIGHIQR_total_costing," ) "),
            median_iqr_averted_cost = str_c(median_avert_costing, " ( ", LOWIQR_avert_costing, "-",
                                            HIGHIQR_avert_costing," ) "),
            median_iqr_total_days = str_c(median_total_days, " ( ", LOWIQR_total_days, "-",
                                          HIGHIQR_total_days," ) "),
            median_iqr_averted_days = str_c(median_avert_days, " ( ", LOWIQR_avert_days, "-",
                                            HIGHIQR_avert_days," ) ")) %>%
    select(!c("median_avert_costing",  "LOWIQR_avert_costing",
              "HIGHIQR_avert_costing", "median_total_costing" ,
              "LOWIQR_total_costing" , "HIGHIQR_total_costing" ,
              "median_avert_days",  "LOWIQR_avert_days",
              "HIGHIQR_avert_days", "median_total_days" ,
              "LOWIQR_total_days" , "HIGHIQR_total_days")) %>%
    
    as.data.table()
  return(x)
}
########### nonadj ################
hospital_cost_l <- list()
for (i in 1:length(nonadj)){
  load(nonadj[i])
  vaccine_output_dt[WHO.Region=="PAHO", WHO.Region := "AMRO"]
  hospital_cost_l[[i]] <- mediqr.reg.syn(vaccine_output_dt)
}

########### adj ################
hospital_cost_l_adj <- list()

hospital_cost_l_adj[[1]] <- mediqr.reg.syn(serotype)
hospital_cost_l_adj[[2]] <- mediqr.reg.syn(serotype.eld)
hospital_cost_l_adj[[3]] <- mediqr.reg.syn(improv)
hospital_cost_l_adj[[4]] <- mediqr.reg.syn(improv.eld)
hospital_cost_l_adj[[5]] <- mediqr.reg.syn(Hib)


hospital_c <- rbindlist(hospital_cost_l)
hospital_cA <- rbindlist(hospital_cost_l_adj)
hospital_c <- rbind(hospital_c,hospital_cA)

hospital_c[, c("vaccine_target_disease","efficacy" ,"coverage"  ,
               "duration", "target disease","target population") := tstrsplit(vaccine_id, "_", fixed=TRUE)]

###!!! not including those to mapped to any vaccine ID
hospital_c <- hospital_c[vaccine_id!="_NA_NA___"]

save(hospital_c,file="outputs/hospitalC_region_averted_syndrome.RData")


######## GLOBAL + ALL LEVELS ############
mediqr.glob.all <- function(vaccine_output_dt){
  x <- vaccine_output_dt%>% group_by(Pathogen, class, Infectious.syndrome,
                                     vaccine_id, run) %>%
    summarise(who_region_cost_A = sum(avertable_cost_cases, na.rm=TRUE),
              who_region_cost_T = sum(cost_cases, na.rm=TRUE),
              who_region_days_A = sum(avertable_days_cases, na.rm=TRUE),
              who_region_days_T = sum(days_cases, na.rm=TRUE))  %>%
    group_by(Pathogen, class, Infectious.syndrome,
                     vaccine_id) %>%
    summarise(median_avert_costing = median(who_region_cost_A, na.rm=TRUE),
              LOWIQR_avert_costing = quantile(who_region_cost_A,0.25, na.rm = TRUE),
              HIGHIQR_avert_costing = quantile(who_region_cost_A,0.75, na.rm = TRUE),
              median_total_costing = median(who_region_cost_T, na.rm=TRUE),
              LOWIQR_total_costing = quantile(who_region_cost_T,0.25, na.rm = TRUE),
              HIGHIQR_total_costing = quantile(who_region_cost_T,0.75, na.rm = TRUE),
              median_avert_days = median(who_region_days_A, na.rm=TRUE),
              LOWIQR_avert_days = quantile(who_region_days_A,0.25, na.rm = TRUE),
              HIGHIQR_avert_days = quantile(who_region_days_A,0.75, na.rm = TRUE),
              median_total_days = median(who_region_days_T, na.rm=TRUE),
              LOWIQR_total_days = quantile(who_region_days_T,0.25, na.rm = TRUE),
              HIGHIQR_total_days = quantile(who_region_days_T,0.75, na.rm = TRUE)) %>%
    mutate(across(median_avert_costing:HIGHIQR_total_days,
                  ~ format(., big.mark = ",", scientific = F)))%>%
    mutate( median_iqr_total_cost = str_c(median_total_costing, " ( ", LOWIQR_total_costing, "-",
                                          HIGHIQR_total_costing," ) "),
            median_iqr_averted_cost = str_c(median_avert_costing, " ( ", LOWIQR_avert_costing, "-",
                                            HIGHIQR_avert_costing," ) "),
            median_iqr_total_days = str_c(median_total_days, " ( ", LOWIQR_total_days, "-",
                                          HIGHIQR_total_days," ) "),
            median_iqr_averted_days = str_c(median_avert_days, " ( ", LOWIQR_avert_days, "-",
                                            HIGHIQR_avert_days," ) ")) %>%
    select(!c("median_avert_costing",  "LOWIQR_avert_costing",
              "HIGHIQR_avert_costing", "median_total_costing" ,
              "LOWIQR_total_costing" , "HIGHIQR_total_costing" ,
              "median_avert_days",  "LOWIQR_avert_days",
              "HIGHIQR_avert_days", "median_total_days" ,
              "LOWIQR_total_days" , "HIGHIQR_total_days")) %>%
    
    as.data.table()
  return(x)
}
########### nonadj ################
hospital_cost_l <- list()
for (i in 1:length(nonadj)){
  load(nonadj[i])
  vaccine_output_dt[WHO.Region=="PAHO", WHO.Region := "AMRO"]
  hospital_cost_l[[i]] <- mediqr.glob.all(vaccine_output_dt)
}

########### adj ################
hospital_cost_l_adj <- list()

hospital_cost_l_adj[[1]] <- mediqr.glob.all(serotype)
hospital_cost_l_adj[[2]] <- mediqr.glob.all(serotype.eld)
hospital_cost_l_adj[[3]] <- mediqr.glob.all(improv)
hospital_cost_l_adj[[4]] <- mediqr.glob.all(improv.eld)
hospital_cost_l_adj[[5]] <- mediqr.glob.all(Hib)


hospital_c <- rbindlist(hospital_cost_l)
hospital_cA <- rbindlist(hospital_cost_l_adj)
hospital_c <- rbind(hospital_c,hospital_cA)

hospital_c[, c("vaccine_target_disease","efficacy" ,"coverage"  ,
               "duration", "target disease","target population") := tstrsplit(vaccine_id, "_", fixed=TRUE)]

###!!! not including those to mapped to any vaccine ID
hospital_c <- hospital_c[vaccine_id!="_NA_NA___"]
hospital_c <- hospital_c[Pathogen!="NA"]

save(hospital_c,file="outputs/hospitalC_global_averted.RData")


######## GLOBAL + CLASS ############

mediqr.glob.class <- function(vaccine_output_dt){
  x <- vaccine_output_dt%>% group_by(class, 
                                             vaccine_id, run) %>%
    summarise(who_region_cost_A = sum(avertable_cost_cases, na.rm=TRUE),
              who_region_cost_T = sum(cost_cases, na.rm=TRUE),
              who_region_days_A = sum(avertable_days_cases, na.rm=TRUE),
              who_region_days_T = sum(days_cases, na.rm=TRUE))  %>%
    group_by(class, 
             vaccine_id) %>%
    summarise(median_avert_costing = median(who_region_cost_A, na.rm=TRUE),
              LOWIQR_avert_costing = quantile(who_region_cost_A,0.25, na.rm = TRUE),
              HIGHIQR_avert_costing = quantile(who_region_cost_A,0.75, na.rm = TRUE),
              median_total_costing = median(who_region_cost_T, na.rm=TRUE),
              LOWIQR_total_costing = quantile(who_region_cost_T,0.25, na.rm = TRUE),
              HIGHIQR_total_costing = quantile(who_region_cost_T,0.75, na.rm = TRUE),
              median_avert_days = median(who_region_days_A, na.rm=TRUE),
              LOWIQR_avert_days = quantile(who_region_days_A,0.25, na.rm = TRUE),
              HIGHIQR_avert_days = quantile(who_region_days_A,0.75, na.rm = TRUE),
              median_total_days = median(who_region_days_T, na.rm=TRUE),
              LOWIQR_total_days = quantile(who_region_days_T,0.25, na.rm = TRUE),
              HIGHIQR_total_days = quantile(who_region_days_T,0.75, na.rm = TRUE)) %>%
    mutate(across(median_avert_costing:HIGHIQR_total_days,
                  ~ format(., big.mark = ",", scientific = F)))%>%
    mutate( median_iqr_total_cost = str_c(median_total_costing, " ( ", LOWIQR_total_costing, "-",
                                          HIGHIQR_total_costing," ) "),
            median_iqr_averted_cost = str_c(median_avert_costing, " ( ", LOWIQR_avert_costing, "-",
                                            HIGHIQR_avert_costing," ) "),
            median_iqr_total_days = str_c(median_total_days, " ( ", LOWIQR_total_days, "-",
                                          HIGHIQR_total_days," ) "),
            median_iqr_averted_days = str_c(median_avert_days, " ( ", LOWIQR_avert_days, "-",
                                            HIGHIQR_avert_days," ) ")) %>%
    select(!c("median_avert_costing",  "LOWIQR_avert_costing",
              "HIGHIQR_avert_costing", "median_total_costing" ,
              "LOWIQR_total_costing" , "HIGHIQR_total_costing" ,
              "median_avert_days",  "LOWIQR_avert_days",
              "HIGHIQR_avert_days", "median_total_days" ,
              "LOWIQR_total_days" , "HIGHIQR_total_days")) %>%
    
    as.data.table()
  return(x)
}
########### nonadj ################
hospital_cost_l <- list()
for (i in 1:length(nonadj)){
  load(nonadj[i])
  vaccine_output_dt[WHO.Region=="PAHO", WHO.Region := "AMRO"]
  hospital_cost_l[[i]] <- mediqr.glob.class(vaccine_output_dt)
}

########### adj ################
hospital_cost_l_adj <- list()

hospital_cost_l_adj[[1]] <- mediqr.glob.class(serotype)
hospital_cost_l_adj[[2]] <- mediqr.glob.class(serotype.eld)
hospital_cost_l_adj[[3]] <- mediqr.glob.class(improv)
hospital_cost_l_adj[[4]] <- mediqr.glob.class(improv.eld)
hospital_cost_l_adj[[5]] <- mediqr.glob.class(Hib)


hospital_c <- rbindlist(hospital_cost_l)
hospital_cA <- rbindlist(hospital_cost_l_adj)
hospital_c <- rbind(hospital_c,hospital_cA)

hospital_c[, c("vaccine_target_disease","efficacy" ,"coverage"  ,
               "duration", "target disease","target population") := tstrsplit(vaccine_id, "_", fixed=TRUE)]

###!!! not including those to mapped to any vaccine ID
hospital_c <- hospital_c[vaccine_id!="_NA_NA___"]

save(hospital_c,file="outputs/hospitalC_global_averted_class.RData")


######## GLOBAL + PATHOGEN ############

mediqr.glob.path <- function(vaccine_output_dt){
  x <- vaccine_output_dt%>% group_by(Pathogen, 
                                     vaccine_id, run) %>%
    summarise(who_region_cost_A = sum(avertable_cost_cases, na.rm=TRUE),
              who_region_cost_T = sum(cost_cases, na.rm=TRUE),
              who_region_days_A = sum(avertable_days_cases, na.rm=TRUE),
              who_region_days_T = sum(days_cases, na.rm=TRUE))  %>%
    group_by(Pathogen, 
             vaccine_id) %>%
    summarise(median_avert_costing = median(who_region_cost_A, na.rm=TRUE),
              LOWIQR_avert_costing = quantile(who_region_cost_A,0.25, na.rm = TRUE),
              HIGHIQR_avert_costing = quantile(who_region_cost_A,0.75, na.rm = TRUE),
              median_total_costing = median(who_region_cost_T, na.rm=TRUE),
              LOWIQR_total_costing = quantile(who_region_cost_T,0.25, na.rm = TRUE),
              HIGHIQR_total_costing = quantile(who_region_cost_T,0.75, na.rm = TRUE),
              median_avert_days = median(who_region_days_A, na.rm=TRUE),
              LOWIQR_avert_days = quantile(who_region_days_A,0.25, na.rm = TRUE),
              HIGHIQR_avert_days = quantile(who_region_days_A,0.75, na.rm = TRUE),
              median_total_days = median(who_region_days_T, na.rm=TRUE),
              LOWIQR_total_days = quantile(who_region_days_T,0.25, na.rm = TRUE),
              HIGHIQR_total_days = quantile(who_region_days_T,0.75, na.rm = TRUE)) %>%
    mutate(across(median_avert_costing:HIGHIQR_total_days,
                  ~ format(., big.mark = ",", scientific = F)))%>%
    mutate( median_iqr_total_cost = str_c(median_total_costing, " ( ", LOWIQR_total_costing, "-",
                                          HIGHIQR_total_costing," ) "),
            median_iqr_averted_cost = str_c(median_avert_costing, " ( ", LOWIQR_avert_costing, "-",
                                            HIGHIQR_avert_costing," ) "),
            median_iqr_total_days = str_c(median_total_days, " ( ", LOWIQR_total_days, "-",
                                          HIGHIQR_total_days," ) "),
            median_iqr_averted_days = str_c(median_avert_days, " ( ", LOWIQR_avert_days, "-",
                                            HIGHIQR_avert_days," ) ")) %>%
    select(!c("median_avert_costing",  "LOWIQR_avert_costing",
              "HIGHIQR_avert_costing", "median_total_costing" ,
              "LOWIQR_total_costing" , "HIGHIQR_total_costing" ,
              "median_avert_days",  "LOWIQR_avert_days",
              "HIGHIQR_avert_days", "median_total_days" ,
              "LOWIQR_total_days" , "HIGHIQR_total_days")) %>%
    
    as.data.table()
  return(x)
}
########### nonadj ################
hospital_cost_l <- list()
for (i in 1:length(nonadj)){
  load(nonadj[i])
  vaccine_output_dt[WHO.Region=="PAHO", WHO.Region := "AMRO"]
  hospital_cost_l[[i]] <- mediqr.glob.path(vaccine_output_dt)
}

########### adj ################
hospital_cost_l_adj <- list()

hospital_cost_l_adj[[1]] <- mediqr.glob.path(serotype)
hospital_cost_l_adj[[2]] <- mediqr.glob.path(serotype.eld)
hospital_cost_l_adj[[3]] <- mediqr.glob.path(improv)
hospital_cost_l_adj[[4]] <- mediqr.glob.path(improv.eld)
hospital_cost_l_adj[[5]] <- mediqr.glob.path(Hib)


hospital_c <- rbindlist(hospital_cost_l)
hospital_cA <- rbindlist(hospital_cost_l_adj)
hospital_c <- rbind(hospital_c,hospital_cA)

hospital_c[, c("vaccine_target_disease","efficacy" ,"coverage"  ,
               "duration", "target disease","target population") := tstrsplit(vaccine_id, "_", fixed=TRUE)]

###!!! not including those to mapped to any vaccine ID
hospital_c <- hospital_c[vaccine_id!="_NA_NA___"]
hospital_c <- hospital_c[Pathogen!="NA"]

save(hospital_c,file="outputs/hospitalC_global_averted_pathogen.RData")

######## GLOBAL + SYNDROME ############

mediqr.glob.syn <- function(vaccine_output_dt){
  x <- vaccine_output_dt%>% group_by( Infectious.syndrome,
                                      vaccine_id, run) %>%
    summarise(who_region_cost_A = sum(avertable_cost_cases, na.rm=TRUE),
              who_region_cost_T = sum(cost_cases, na.rm=TRUE),
              who_region_days_A = sum(avertable_days_cases, na.rm=TRUE),
              who_region_days_T = sum(days_cases, na.rm=TRUE))  %>%
    group_by( Infectious.syndrome,
              vaccine_id) %>%
    summarise(median_avert_costing = median(who_region_cost_A, na.rm=TRUE),
              LOWIQR_avert_costing = quantile(who_region_cost_A,0.25, na.rm = TRUE),
              HIGHIQR_avert_costing = quantile(who_region_cost_A,0.75, na.rm = TRUE),
              median_total_costing = median(who_region_cost_T, na.rm=TRUE),
              LOWIQR_total_costing = quantile(who_region_cost_T,0.25, na.rm = TRUE),
              HIGHIQR_total_costing = quantile(who_region_cost_T,0.75, na.rm = TRUE),
              median_avert_days = median(who_region_days_A, na.rm=TRUE),
              LOWIQR_avert_days = quantile(who_region_days_A,0.25, na.rm = TRUE),
              HIGHIQR_avert_days = quantile(who_region_days_A,0.75, na.rm = TRUE),
              median_total_days = median(who_region_days_T, na.rm=TRUE),
              LOWIQR_total_days = quantile(who_region_days_T,0.25, na.rm = TRUE),
              HIGHIQR_total_days = quantile(who_region_days_T,0.75, na.rm = TRUE)) %>%
    mutate(across(median_avert_costing:HIGHIQR_total_days,
                  ~ format(., big.mark = ",", scientific = F)))%>%
    mutate( median_iqr_total_cost = str_c(median_total_costing, " ( ", LOWIQR_total_costing, "-",
                                          HIGHIQR_total_costing," ) "),
            median_iqr_averted_cost = str_c(median_avert_costing, " ( ", LOWIQR_avert_costing, "-",
                                            HIGHIQR_avert_costing," ) "),
            median_iqr_total_days = str_c(median_total_days, " ( ", LOWIQR_total_days, "-",
                                          HIGHIQR_total_days," ) "),
            median_iqr_averted_days = str_c(median_avert_days, " ( ", LOWIQR_avert_days, "-",
                                            HIGHIQR_avert_days," ) ")) %>%
    select(!c("median_avert_costing",  "LOWIQR_avert_costing",
              "HIGHIQR_avert_costing", "median_total_costing" ,
              "LOWIQR_total_costing" , "HIGHIQR_total_costing" ,
              "median_avert_days",  "LOWIQR_avert_days",
              "HIGHIQR_avert_days", "median_total_days" ,
              "LOWIQR_total_days" , "HIGHIQR_total_days")) %>%
    
    as.data.table()
  return(x)
}
########### nonadj ################
hospital_cost_l <- list()
for (i in 1:length(nonadj)){
  load(nonadj[i])
  vaccine_output_dt[WHO.Region=="PAHO", WHO.Region := "AMRO"]
  hospital_cost_l[[i]] <- mediqr.glob.syn(vaccine_output_dt)
}

########### adj ################
hospital_cost_l_adj <- list()

hospital_cost_l_adj[[1]] <- mediqr.glob.syn(serotype)
hospital_cost_l_adj[[2]] <- mediqr.glob.syn(serotype.eld)
hospital_cost_l_adj[[3]] <- mediqr.glob.syn(improv)
hospital_cost_l_adj[[4]] <- mediqr.glob.syn(improv.eld)
hospital_cost_l_adj[[5]] <- mediqr.glob.syn(Hib)


hospital_c <- rbindlist(hospital_cost_l)
hospital_cA <- rbindlist(hospital_cost_l_adj)
hospital_c <- rbind(hospital_c,hospital_cA)

hospital_c[, c("vaccine_target_disease","efficacy" ,"coverage"  ,
               "duration", "target disease","target population") := tstrsplit(vaccine_id, "_", fixed=TRUE)]

###!!! not including those to mapped to any vaccine ID
hospital_c <- hospital_c[vaccine_id!="_NA_NA___"]

save(hospital_c,file="outputs/hospitalC_global_averted_syndrome.RData")


######## GLOBAL + PATHOGEN + SYNDROME ############
mediqr.glob.syn.path <- function(vaccine_output_dt){
  x <- vaccine_output_dt%>% group_by(Pathogen,Infectious.syndrome,
                                     vaccine_id, run) %>%
    summarise(who_region_cost_A = sum(avertable_cost_cases, na.rm=TRUE),
              who_region_cost_T = sum(cost_cases, na.rm=TRUE),
              who_region_days_A = sum(avertable_days_cases, na.rm=TRUE),
              who_region_days_T = sum(days_cases, na.rm=TRUE))  %>%
    group_by(Pathogen,Infectious.syndrome,
             vaccine_id) %>%
    summarise(median_avert_costing = median(who_region_cost_A, na.rm=TRUE),
              LOWIQR_avert_costing = quantile(who_region_cost_A,0.25, na.rm = TRUE),
              HIGHIQR_avert_costing = quantile(who_region_cost_A,0.75, na.rm = TRUE),
              median_total_costing = median(who_region_cost_T, na.rm=TRUE),
              LOWIQR_total_costing = quantile(who_region_cost_T,0.25, na.rm = TRUE),
              HIGHIQR_total_costing = quantile(who_region_cost_T,0.75, na.rm = TRUE),
              median_avert_days = median(who_region_days_A, na.rm=TRUE),
              LOWIQR_avert_days = quantile(who_region_days_A,0.25, na.rm = TRUE),
              HIGHIQR_avert_days = quantile(who_region_days_A,0.75, na.rm = TRUE),
              median_total_days = median(who_region_days_T, na.rm=TRUE),
              LOWIQR_total_days = quantile(who_region_days_T,0.25, na.rm = TRUE),
              HIGHIQR_total_days = quantile(who_region_days_T,0.75, na.rm = TRUE)) %>%
    mutate(across(median_avert_costing:HIGHIQR_total_days,
                  ~ format(., big.mark = ",", scientific = F)))%>%
    mutate( median_iqr_total_cost = str_c(median_total_costing, " ( ", LOWIQR_total_costing, "-",
                                          HIGHIQR_total_costing," ) "),
            median_iqr_averted_cost = str_c(median_avert_costing, " ( ", LOWIQR_avert_costing, "-",
                                            HIGHIQR_avert_costing," ) "),
            median_iqr_total_days = str_c(median_total_days, " ( ", LOWIQR_total_days, "-",
                                          HIGHIQR_total_days," ) "),
            median_iqr_averted_days = str_c(median_avert_days, " ( ", LOWIQR_avert_days, "-",
                                            HIGHIQR_avert_days," ) ")) %>%
    select(!c("median_avert_costing",  "LOWIQR_avert_costing",
              "HIGHIQR_avert_costing", "median_total_costing" ,
              "LOWIQR_total_costing" , "HIGHIQR_total_costing" ,
              "median_avert_days",  "LOWIQR_avert_days",
              "HIGHIQR_avert_days", "median_total_days" ,
              "LOWIQR_total_days" , "HIGHIQR_total_days")) %>%
    
    as.data.table()
  return(x)
}
########### nonadj ################
hospital_cost_l <- list()
for (i in 1:length(nonadj)){
  load(nonadj[i])
  vaccine_output_dt[WHO.Region=="PAHO", WHO.Region := "AMRO"]
  hospital_cost_l[[i]] <- mediqr.glob.syn.path(vaccine_output_dt)
}

########### adj ################
hospital_cost_l_adj <- list()

hospital_cost_l_adj[[1]] <- mediqr.glob.syn.path(serotype)
hospital_cost_l_adj[[2]] <- mediqr.glob.syn.path(serotype.eld)
hospital_cost_l_adj[[3]] <- mediqr.glob.syn.path(improv)
hospital_cost_l_adj[[4]] <- mediqr.glob.syn.path(improv.eld)
hospital_cost_l_adj[[5]] <- mediqr.glob.syn.path(Hib)


hospital_c <- rbindlist(hospital_cost_l)
hospital_cA <- rbindlist(hospital_cost_l_adj)
hospital_c <- rbind(hospital_c,hospital_cA)

hospital_c[, c("vaccine_target_disease","efficacy" ,"coverage"  ,
               "duration", "target disease","target population") := tstrsplit(vaccine_id, "_", fixed=TRUE)]

###!!! not including those to mapped to any vaccine ID
hospital_c <- hospital_c[vaccine_id!="_NA_NA___"]
hospital_c <- hospital_c[Pathogen!="NA"]

save(hospital_c,file="outputs/hospitalC_global_averted_pathogen_syndrome.RData")


######## GLOBAL + UNIT ############
hospital_c_l <- list()

for (i in 1:length(myfiles)){
  
  load(myfiles[i])
  vaccine_output_dt[WHO.Region=="PAHO", WHO.Region := "AMRO"]
  hospitalC_region_averted <- vaccine_output_dt %>% 
    filter(!is.na(Pathogen)& vaccine_id!="_NA_NA___") %>%
    group_by(iso3c, Pathogen, class, Infectious.syndrome, run) %>%
    slice(1) %>%
    group_by(Pathogen, class, Infectious.syndrome) %>%
    summarise(Median_unitcost = median(los.cost, na.rm=TRUE),
              LOWIQR_unitcost = quantile(los.cost,0.25, na.rm = TRUE),
              HIGIQR_unitcost = quantile(los.cost,0.75, na.rm = TRUE),
              Median_los = median(los.DRI, na.rm=TRUE),
              LOWIQR_los = quantile(los.DRI,0.25, na.rm = TRUE),
              HIGIQR_los = quantile(los.DRI,0.75, na.rm = TRUE),)%>%
    as.data.table()
  hospital_c_l[[i]] <- hospitalC_region_averted
}


hospital_c<- rbindlist(hospital_c_l)
save(hospital_c,file="outputs/hospitalC_global_los.RData")
rm(hospital_c)
rm(hospitalC_region_averted)
rm(hospital_c_l)

###### !!! problem is this isn't population weighted & previous unit costs are
### produce country level averages, then weight using population like for AMR-UCR code:

######## GLOBAL + UNIT - population weighted ############
hospital_c_l <- list()

N <- as.data.table(read.csv("data_all/Population-EstimatesData_092020.csv"))

N <- N[Indicator.Code=="SP.POP.TOTL"] ## total population

### keep 2019 values
N <- N[ , c("Country.Code","X2019")]
setnames(N, "X2019", "npop")

for (i in 1:length(myfiles)){
  
  load(myfiles[i])
  
  vaccine_output_dt[WHO.Region=="PAHO", WHO.Region := "AMRO"]
 x <-  merge(vaccine_output_dt,N, by.x="iso3c",by.y="Country.Code",all.x=TRUE,all.y=FALSE)
  hospitalC_region_averted <- x %>% 
    filter(!is.na(Pathogen) & !is.na(npop) & vaccine_id!="_NA_NA___") %>%
    group_by(Pathogen, class, Infectious.syndrome,run) %>%
    summarise(weighted_costing_los = weighted.mean(los.cost, npop)) %>%
    group_by(Pathogen, class, Infectious.syndrome) %>%
    summarise(Median_unitcost = median(weighted_costing_los, na.rm=TRUE),
              LOWIQR_unitcost = quantile(weighted_costing_los,0.25, na.rm = TRUE),
              HIGIQR_unitcost = quantile(weighted_costing_los,0.75, na.rm = TRUE)) %>%
    as.data.table()
  hospital_c_l[[i]] <- hospitalC_region_averted
}


hospital_c<- rbindlist(hospital_c_l)
save(hospital_c,file="outputs/hospitalC_global_los_npop.RData")
rm(hospital_c)
rm(hospitalC_region_averted)
rm(hospital_c_l)


######## GLOBAL Hospital Costs to get 1 final big number ############

mediqr.glob <- function(vaccine_output_dt){
  x <-   vaccine_output_dt[WHO.Region=="PAHO", WHO.Region := "AMRO"]
  x <- x %>% 
    filter(!is.na(Pathogen)& vaccine_id!="_NA_NA___") %>%
    group_by(vaccine_id, Infectious.syndrome, run) %>%
    summarise(hospital_cost =sum(cost_cases, na.rm=TRUE),
              avert_hospital_cost=sum(avertable_cost_cases,na.rm=TRUE),
              hospital_days=sum(days_cases, na.rm=TRUE),
              avert_hospital_days=sum(avertable_days_cases, na.rm=TRUE))%>%
    as.data.table()
  return(x)
}
########### nonadj ################
hospital_cost_l <- list()
for (i in 1:length(nonadj)){
  load(nonadj[i])
  hospital_cost_l[[i]] <- mediqr.glob(vaccine_output_dt)
}

########### adj ################
hospital_cost_l_adj <- list()

hospital_cost_l_adj[[1]] <- mediqr.glob(serotype)
hospital_cost_l_adj[[2]] <- mediqr.glob(serotype.eld)
hospital_cost_l_adj[[3]] <- mediqr.glob(improv)
hospital_cost_l_adj[[4]] <- mediqr.glob(improv.eld)
hospital_cost_l_adj[[5]] <- mediqr.glob(Hib)

hospital_c <- rbindlist(hospital_cost_l)
rm(hospital_cost_l)
hospital_cA <- rbindlist(hospital_cost_l_adj)
rm(hospital_cost_l_adj)

hospital_c <- rbind(hospital_c,hospital_cA)

hospital_c[, c("vaccine_target_disease","efficacy" ,"coverage"  ,
               "duration", "target disease","target population") := tstrsplit(vaccine_id, "_", fixed=TRUE)]

###!!! not including those to mapped to any vaccine ID
hospital_c <- hospital_c[vaccine_id!="_NA_NA___"]


save(hospital_c,file="outputs/total_global_cost.RData")

#### total class numbers #####
load("outputs/hospitalC_global_averted_class.RData")

class.cost <- hospital_c[,c("class","median_iqr_total_cost")]
class.cost <- distinct(class.cost)

## extract median values
### remove commas from numeric strings
class.cost[ , median_iqr_total_cost := gsub(",", "", gsub("([a-zA-Z]),", "\\1 ", median_iqr_total_cost)) ]

### separate out numbers in/across brackets
Separate <- function(...) separate(..., sep = "[^[:alnum:].]+", convert = TRUE)
class.cost <- class.cost %>%
  Separate(median_iqr_total_cost, into = c("cost_med", "cost_lo", "cost_hi", NA)) %>%
  group_by(class) %>%
  summarise(total = sum(cost_med))

######## USA and Thailand ############

mediqr.x <- function(vaccine_output_dt){
  
 x <-  vaccine_output_dt[iso3c=="USA"|iso3c=="THA"]
  x <- x %>% group_by(iso3c, Pathogen,
                                      vaccine_id, run) %>%
    summarise(who_region_cost_A = sum(avertable_cost_cases, na.rm=TRUE),
              who_region_cost_T = sum(cost_cases, na.rm=TRUE),
              who_region_days_A = sum(avertable_days_cases, na.rm=TRUE),
              who_region_days_T = sum(days_cases, na.rm=TRUE))  %>%
    group_by( iso3c,
              vaccine_id, Pathogen) %>%
    summarise(median_avert_costing = median(who_region_cost_A, na.rm=TRUE),
              LOWIQR_avert_costing = quantile(who_region_cost_A,0.25, na.rm = TRUE),
              HIGHIQR_avert_costing = quantile(who_region_cost_A,0.75, na.rm = TRUE),
              median_total_costing = median(who_region_cost_T, na.rm=TRUE),
              LOWIQR_total_costing = quantile(who_region_cost_T,0.25, na.rm = TRUE),
              HIGHIQR_total_costing = quantile(who_region_cost_T,0.75, na.rm = TRUE),
              median_avert_days = median(who_region_days_A, na.rm=TRUE),
              LOWIQR_avert_days = quantile(who_region_days_A,0.25, na.rm = TRUE),
              HIGHIQR_avert_days = quantile(who_region_days_A,0.75, na.rm = TRUE),
              median_total_days = median(who_region_days_T, na.rm=TRUE),
              LOWIQR_total_days = quantile(who_region_days_T,0.25, na.rm = TRUE),
              HIGHIQR_total_days = quantile(who_region_days_T,0.75, na.rm = TRUE)) %>%
    
    as.data.table()
  return(x)
}
########### nonadj ################
hospital_cost_l <- list()
for (i in 1:length(nonadj)){
  load(nonadj[i])
  vaccine_output_dt[WHO.Region=="PAHO", WHO.Region := "AMRO"]
  hospital_cost_l[[i]] <- mediqr.x(vaccine_output_dt)
}

########### adj ################
hospital_cost_l_adj <- list()

hospital_cost_l_adj[[1]] <- mediqr.x(serotype)
hospital_cost_l_adj[[2]] <- mediqr.x(serotype.eld)
hospital_cost_l_adj[[3]] <- mediqr.x(improv)
hospital_cost_l_adj[[4]] <- mediqr.x(improv.eld)
hospital_cost_l_adj[[5]] <- mediqr.x(Hib)


hospital_c <- rbindlist(hospital_cost_l)
hospital_cA <- rbindlist(hospital_cost_l_adj)
hospital_c <- rbind(hospital_c,hospital_cA)

hospital_c[, c("vaccine_target_disease","efficacy" ,"coverage"  ,
               "duration", "target disease","target population") := tstrsplit(vaccine_id, "_", fixed=TRUE)]

keep_vac <- c(               
  "Haemophilus influenzae type B_both_0.9_5 years_All_6, 10, 14 weeks"                                                           
  ,"Salmonella Typhi_0.85_0.7_15 years_All_9 months"                                                                              
  , "Staphylococcus aureus_0.6_0.7_5 years_All_All age groups"                                                               
  , "E. coli - non-diarrhogenic_0.7_0.7_5 years_All except Diarrhoea_All age groups"                                      
  , "ETEC_0.6_0.7_5 years_Diarrhoea_6 months"                                                                   
  , "Group A streptococcus_0.7_0.7_5 years_All_6 weeks"                                                                            
  , "Klebsiella pneumoniae - all_0.7_0.7_5 years_All_All age groups"                                                      
  , "Streptococcus pneumoniae - Improved_both_0.9_5 years_BSI, CNS infections, Cardiac infections, LRI_6 weeks & elderly age group"
  , "Pseudomonas aeruginosa_0.7_0.7_5 years_BSI, LRI and thorax infections_All age groups"                                         
  ,"Enterococcus faecium_0.7_0.7_5 years_All_All age groups"                                                                      
  , "Salmonella paratyphi_0.7_0.7_5 years_All_9 months"                                                                            
  , "Acinetobacter baumannii - all_0.7_0.7_5 years_All_All age groups"                                                            
  , "Mycobacterium tuberculosis - Improved_0.8_0.7_10 years_All_0 weeks + boost every 10 years"                                    
  , "Shigella_0.6_0.7_5 years_All_6 months" 
  ,"Non-typhoidal Salmonella_0.8_0.7_5 years_All_6 weeks & 9 months" ) 

  global_hospital <- hospital_c[hospital_c$vaccine_id %in% keep_vac]
  global_hospital[vaccine_id=="ETEC_0.6_0.7_5 years_Diarrhoea_6 months",
                  median_total_costing :=0]

  tha <- global_hospital[iso3c=="THA"]
  usa <- global_hospital[iso3c=="USA"]
  
  sum(tha$median_total_costing)
  sum(usa$median_total_costing)
