==> 6b_journalplots.R <==
############ THIS SCRIPT IS A COMPILATION OF OTHER SCRIPTS WRITTEN DURING
########### THE MANUSCRIPT SUBMISSION PROCESS

############# duplication likely as copied over from 3 different editing rounds to keep in 1 place

#### be careful running as does save data sets and plots (latest figure 4 & 5)
#### although some also currently hashed-out to avoid overwriting


############## CREATING PLOTS FROM OUTPUTS ##############
rm(list=ls())

#### loading packages
library(tidyr)
library(data.table)
library(dplyr)
library(stringr)
library(ggplot2)
library(patchwork) ## added in journal resub phase

#### !!! note throughout did last minute changes to change e.g. AFRO labelling to AFR
### ideally this would be dealt with at start of analysis

######### cleaning functions ################

keep_vac <- c(               
  "Haemophilus influenzae type B_both_0.9_5 years_All_6, 10, 14 weeks"                                                           
  ,"Salmonella Typhi_0.85_0.7_15 years_All_9 months"                                                                              
  , "Staphylococcus aureus_0.6_0.7_5 years_All_All age groups"                                                               
  , "E. coli - non-diarrhogenic_0.7_0.7_5 years_All except Diarrhoea_All age groups"                                      
  , "ETEC_0.6_0.7_5 years_Diarrhoea_6 months"                                                                   
  , "Group A streptococcus_0.7_0.7_5 years_All_6 weeks"                                                                            
  , "Klebsiella pneumoniae - all_0.7_0.7_5 years_All_All age groups"                                                      
  , "Streptococcus pneumoniae - Improved_both_0.9_5 years_BSI, CNS infections, Cardiac infections, LRI_6 weeks & elderly age group"
  , "Pseudomonas aeruginosa_0.7_0.7_5 years_BSI, LRI and thorax infections_All age groups"                                         
  ,"Enterococcus faecium_0.7_0.7_5 years_All_All age groups"                                                                      
  , "Salmonella paratyphi_0.7_0.7_5 years_All_9 months"                                                                            
  , "Acinetobacter baumannii - all_0.7_0.7_5 years_All_All age groups"                                                            
  , "Mycobacterium tuberculosis - Improved_0.8_0.7_10 years_All_0 weeks + boost every 10 years"                                    
  , "Shigella_0.6_0.7_5 years_All_6 months"
  ,"Non-typhoidal Salmonella_0.8_0.7_5 years_All_6 weeks & 9 months" ) 
### need to specify different ones for productivity as didn't group
## vaccine scenarios like had done for 

## selecting ones that have the theoretical biggest impact (e.g. most age groups/syndromes)

keep_vac_prod <- c(               
  "Haemophilus influenzae type B_0.93_0.9_5 years_All_6, 10, 14 weeks"  
  ,"Haemophilus influenzae type B_0.69_0.9_5 years_All_6, 10, 14 weeks" 
  ,"Streptococcus pneumoniae - Improved_0.7_0.9_5 years_BSI, CNS infections, Cardiac infections, LRI_6 weeks & elderly age group"
  ,"Streptococcus pneumoniae - Improved_0.5_0.9_5 years_BSI, CNS infections, Cardiac infections, LRI_6 weeks & elderly age group"
  ,"Salmonella Typhi_0.85_0.7_15 years_All_9 months"                                                                              
  , "Staphylococcus aureus_0.6_0.7_5 years_All_All age groups"                                                               
  , "E. coli - non-diarrhogenic_0.7_0.7_5 years_All except Diarrhoea_All age groups"                                      
  , "ETEC_0.6_0.7_5 years_Diarrhoea_6 months"                                                                   
  , "Group A streptococcus_0.7_0.7_5 years_All_6 weeks"                                                                            
  , "Klebsiella pneumoniae - all_0.7_0.7_5 years_All_All age groups"                                                      
  , "Pseudomonas aeruginosa_0.7_0.7_5 years_BSI, LRI and thorax infections_All age groups"                                         
  ,"Enterococcus faecium_0.7_0.7_5 years_All_All age groups"                                                                      
  , "Salmonella paratyphi_0.7_0.7_5 years_All_9 months"                                                                            
  , "Acinetobacter baumannii - all_0.7_0.7_5 years_All_All age groups"                                                            
  , "Mycobacterium tuberculosis - Improved_0.8_0.7_10 years_All_0 weeks + boost every 10 years"                                    
  , "Shigella_0.6_0.7_5 years_All_6 months"
  ,"Non-typhoidal Salmonella_0.8_0.7_5 years_All_6 weeks & 9 months" ) 


cleaning.dt <- function(hospital_c){
  global_hospital <- hospital_c[hospital_c$vaccine_id %in% keep_vac]
  
  ### need to specify for E.coli target syndromes so no double counting
  global_hospital[ , flag := 0]
  global_hospital[(vaccine_id=="ETEC_0.6_0.7_5 years_Diarrhoea_6 months" & 
                     Infectious.syndrome!="Diarrhoea")|(
                       vaccine_id=="E. coli - non-diarrhogenic_0.7_0.7_5 years_All except Diarrhoea_All age groups"& 
                         Infectious.syndrome=="Diarrhoea"),
                  flag := 1]
  global_hospital <-  global_hospital[flag==0]
  global_hospital <-  global_hospital[ ,-c("flag")]
  
  
  return(global_hospital)
}

cleaning.dt.prod <- function(hospital_c){
  global_hospital <- hospital_c[hospital_c$vaccine_id %in% keep_vac_prod]
  
  ### need to specify for E.coli target syndromes so no double counting
  global_hospital[ , flag := 0]
  global_hospital[(vaccine_id=="ETEC_0.6_0.7_5 years_Diarrhoea_6 months" & 
                     Infectious.syndrome!="Diarrhoea")|(
                       vaccine_id=="E. coli - non-diarrhogenic_0.7_0.7_5 years_All except Diarrhoea_All age groups"& 
                         Infectious.syndrome=="Diarrhoea"),
                  flag := 1]
  global_hospital <-  global_hospital[flag==0]
  global_hospital <-  global_hospital[ ,-c("flag")]
  
  global_hospital[Infectious.syndrome=="BSI", Infectious.syndrome := "Bloodstream infections"]
  global_hospital[Infectious.syndrome=="UTI", Infectious.syndrome := "Urinary tract infections"]
  global_hospital[Infectious.syndrome=="TB", Infectious.syndrome := "Tuberculosis"]
  global_hospital[Antibiotic.class=="Methicillin", Antibiotic.class := "Penicillin"]
  global_hospital[Antibiotic.class=="Multi-drug resistance excluding extensive drug resistance in TB",
                  Antibiotic.class := "MDR"]
  global_hospital[Antibiotic.class=="Multi-drug resistance in Salmonella Typhi and Paratyphi",
                  Antibiotic.class := "MDR"]
  
  ### change vaccine ids for those we want to combine
  global_hospital[vaccine_id=="Haemophilus influenzae type B_0.93_0.9_5 years_All_6, 10, 14 weeks"|
                    vaccine_id=="Haemophilus influenzae type B_0.69_0.9_5 years_All_6, 10, 14 weeks" ,
                  vaccine_id := "Haemophilus influenzae type B_both_0.9_5 years_All_6, 10, 14 weeks" ]
  
  global_hospital[vaccine_id=="Streptococcus pneumoniae - Improved_0.7_0.9_5 years_BSI, CNS infections, Cardiac infections, LRI_6 weeks & elderly age group"|
                    vaccine_id=="Streptococcus pneumoniae - Improved_0.5_0.9_5 years_BSI, CNS infections, Cardiac infections, LRI_6 weeks & elderly age group",
                  vaccine_id := "Streptococcus pneumoniae - Improved_both_0.9_5 years_BSI, CNS infections, Cardiac infections, LRI_6 weeks & elderly age group"]
  
  return(global_hospital)
}

getting_numbers_bk <- function(regional_hospital_totals){
  ### remove commas from numeric strings
  regional_hospital_totals[ , median_iqr_total_cost := gsub(",", "", gsub("([a-zA-Z]),", "\\1 ", median_iqr_total_cost)) ]
  regional_hospital_totals[ , median_iqr_averted_cost := gsub(",", "", gsub("([a-zA-Z]),", "\\1 ", median_iqr_averted_cost)) ]
  regional_hospital_totals[ , median_iqr_total_days := gsub(",", "", gsub("([a-zA-Z]),", "\\1 ", median_iqr_total_days)) ]
  regional_hospital_totals[ , median_iqr_averted_days := gsub(",", "", gsub("([a-zA-Z]),", "\\1 ", median_iqr_averted_days)) ]
  
  ### separate out numbers in/across brackets
  Separate <- function(...) separate(..., sep = "[^[:alnum:].]+", convert = TRUE)
  
  regional_hospital_totals <- regional_hospital_totals %>%
    Separate(median_iqr_total_cost, into = c("cost_med", "cost_lo", "cost_hi", NA)) %>%
    Separate(median_iqr_averted_cost, into = c("avert_cost_med", "avert_cost_lo", "avert_cost_hi", NA)) %>%
    Separate(median_iqr_total_days, into = c("days_med", "days_lo", "days_hi", NA)) %>%
    Separate(median_iqr_averted_days, into = c("avert_days_med", "avert_days_lo", "avert_days_hi", NA)) %>%
    as.data.table()
  return(regional_hospital_totals)
}



### formatting ###
options(scipen=10000) ## turn off scientific notation
`%!in%` = Negate(`%in%`) ## create function for negating %in%



###############################################################
#### newer plots #####
#### code for plotting final outputs by Rhys Kingston 2023
### adapted for this repo


if (!require("pacman")) install.packages("pacman")

pacman::p_load(
  "tidyverse"
  , "readr"
  , "here"
  , "janitor"
  , "scales"
  , "rgdal"
  , "sf"
  , "countrycode"
  , "maps"
  , "rgeos"
  , "devtools"
  , "treemapify"
  , "ggbeeswarm"
  , "viridis"
  , "plotly"
)

########## loading data
load("outputs/regional_hospital_4additionalplots.RData") ## adding sc2 for productivity
load("outputs/prod_additional4plots_sc2.RData")

regional_prod_deaths <- all.prod.methods[, lapply(.SD, sum, na.rm=TRUE),
                                         by = c(".id","Pathogen","vaccine_id"),
                                         .SDcols=c("sc1_total","sc2_total",
                                                   "sc1_averted","sc2_averted")]


### merge together
all.dt <- merge(regional_prod_deaths, regional_hospital_totals, by.x=c(".id","Pathogen","vaccine_id"),
                by.y=c("WHO.Region","Pathogen","vaccine_id"))
### !! note that the total costing set to 0 for ETEC is just because all e.coli is already costed (included diarrheoa)
## in the EXPEC - so applicable to both but removed to remove double counting


## sum across vaccines for pathogens
all.dt <- all.dt[, lapply(.SD, sum, na.rm=TRUE),
                 by = c("Pathogen",
                        ".id"),
                 .SDcols=c(    "sc1_total","sc2_total",
                               "sc1_averted","sc2_averted" ,         
                               "avert_cost_med"  ,"avert_cost_lo",
                               "avert_cost_hi" ,
                               "cost_med",  "cost_lo" ,
                               "cost_hi", "avert_days_med"    ,
                               "avert_days_lo"  ,   "avert_days_hi" ,
                               "days_med"  ,   "days_lo"   , 
                               "days_hi"  )]

all.dt.temp <- all.dt[,.(median_cost_sum_G=sum(cost_med)),
                      by=c("Pathogen")]

all.dt <- merge(all.dt, all.dt.temp, by=c("Pathogen"),
                allow.cartesian = TRUE)

all.dt.temp <- all.dt[,.(median_prod_cost_sum_G=sum(sc1_total)),
                      by=c("Pathogen")]

all.dt <- merge(all.dt, all.dt.temp, by=c("Pathogen"),
                allow.cartesian = TRUE)

all.dt[ , global_total := median_prod_cost_sum_G+median_cost_sum_G]

##### Line plot for hospital costs
options(scipen=10000) ## turn off scientific notation

########### total costs ####################

### renaming regions
all.dt[.id=="AFRO", .id:="AFR"]
all.dt[.id=="AMRO", .id:="AMR"]
all.dt[.id=="EMRO", .id:="EMR"]
all.dt[.id=="EURO", .id:="EUR"]
all.dt[.id=="SEARO", .id:="SEAR"]
all.dt[.id=="WPRO", .id:="WPR"]



load("outputs/working_life_years_lost.Rdata")

output_hc <- as.data.table(output_hc)
### getting one vaccine/one DRI per group
output_hc <- cleaning.dt.prod(output_hc)

regional_prod_path  <-output_hc[, lapply(.SD, sum, na.rm=TRUE),
                                by = c("Pathogen",
                                       ".id"),
                                .SDcols=c("averted_HC")]
regional_prod_path[.id=="PAHO",.id := "AMRO"]
regional_prod_path[.id=="AFRO", .id:="AFR"]
regional_prod_path[.id=="AMRO", .id:="AMR"]
regional_prod_path[.id=="EMRO", .id:="EMR"]
regional_prod_path[.id=="EURO", .id:="EUR"]
regional_prod_path[.id=="SEARO", .id:="SEAR"]
regional_prod_path[.id=="WPRO", .id:="WPR"]

all.dt2 <- merge(regional_prod_path,all.dt, by=c("Pathogen",".id"))

### transform dataset
hospital_avert <- all.dt2[ ,c("Pathogen",
                              ".id",
                              "avert_cost_med"  ,      
                              "avert_cost_lo",
                              "avert_cost_hi",
                              "avert_days_med"  ,
                              "avert_days_lo" ,
                              "avert_days_hi" )]

transf_fact <- max(hospital_avert$avert_cost_med)/max(hospital_avert$avert_days_med)

## relabel regions
hospital_avert[.id=="AFRO" , .id := "AFR"]
hospital_avert[.id=="AMRO" , .id := "AMR"]
hospital_avert[.id=="EMRO" , .id := "EMR"]
hospital_avert[.id=="EURO" , .id := "EUR"]
hospital_avert[.id=="SEARO" , .id := "SEAR"]
hospital_avert[.id=="WPRO" , .id := "WPR"]


f5_1 <- ggplot(hospital_avert, aes(x = Pathogen, y = avert_cost_med)) +
  geom_col(fill="lightblue1") +
  geom_errorbar(aes(ymin = avert_cost_lo, ymax = avert_cost_hi),
                position = position_dodge(0.9), width = .2, color='midnightblue')+
  geom_point(aes(y = transf_fact * avert_days_med), size=2, color='darkslategrey')+
  facet_wrap(~.id)+ 
  ggtitle("(A) Hospital Cost Burden") +
  scale_y_continuous(labels = scales::label_number(accuracy = 0.1, scale_cut = scales::cut_short_scale()),
                     sec.axis = sec_axis(trans = ~ . / transf_fact, 
                                         name = "Total Averted Bed Days (for Points)",
                                         labels = scales::label_number(accuracy = 0.1, scale_cut = scales::cut_short_scale())))+
  theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1),
        text = element_text(size = 15), 
        plot.title = element_text(
          face = "bold"
          , family = "sans"
          , size = 20
          , colour = "black",
          hjust = 0.15
        ))+
  ylab("Total Averted Hospital Cost (for Bars)") 


### create min and max values
all.dt[sc1_averted<sc2_averted, min_prod := sc1_averted]
all.dt[sc2_averted<sc1_averted, min_prod := sc2_averted]
all.dt[sc1_averted<sc2_averted, max_prod := sc2_averted]
all.dt[sc2_averted<sc1_averted, max_prod := sc1_averted]

f5_2 <- all.dt %>%
  ggplot() +
  geom_pointrange(
    mapping = aes(
      y = fct_reorder(Pathogen, median_cost_sum_G)
      , x = sc1_averted
      , xmin= min_prod
      , xmax= max_prod
      , colour = .id
    )
    , position = position_dodge(width = 0.5)
    , size = 0.7
    , lwd = 0.7
  ) +
  scale_colour_brewer(palette = "Dark2") +
  scale_x_log10(
    labels = scales::dollar_format(scale = 1e-6, suffix = "mn")
    , position = "top"
  ) +
  labs(
    title = "(B) Averted Labour Productivity Losses ($)",
    y = "Pathogen"
  ) +
  theme_minimal() +
  theme(
    panel.grid.minor.x = element_blank()
    , panel.grid.major.y = element_blank()
    , axis.line.y = element_line(
      colour = "black"
      , linewidth = 1
    )
    , axis.text.y = element_text(
      face = "bold.italic"
      , family = "sans"
      , size = 14
      , colour = "black"
    )
    , axis.text.x.top = element_text(
      family = "sans"
      , size = 12
      , margin = margin(
        b = 10
      )
    )
    , axis.title.x = element_blank()
    , legend.position = c(0.20, 0.98)
    , legend.title = element_blank()
    , plot.title = element_text(
      face = "bold"
      , family = "sans"
      , size = 20
      , colour = "black"
    )
    , plot.subtitle = element_text(
      face = "bold"
      , family = "sans"
      , size = 16
      , colour = "grey50"
    )
  ) +
  guides(
    alpha = "none"
    , colour = guide_legend(
      override.aes = list(
        size = 2
      )
      , nrow = 1))+
  theme(panel.grid.major.y =element_line(colour="black"),
        plot.margin=margin(10,50,10,10)
  )



f5_3 <- all.dt2 %>%
  ggplot() +
  geom_point(
    mapping = aes(
      y = fct_reorder(Pathogen, median_cost_sum_G)
      , x = averted_HC
      , colour = .id
    )
    , position = position_dodge(width = 0.5)
    , size = 7
    , lwd = 0.7
  ) +
  scale_colour_brewer(palette = "Dark2") +
  scale_x_log10(
    labels = unit_format(unit = "mn", scale = 1e-6)
    , position = "top"
  ) +
  labs(
    title = "(C) Averted Working-Life-Years-Lost",
    y = "Pathogen"
  ) +
  theme_minimal() +
  theme(
    panel.grid.minor.x = element_blank()
    , panel.grid.major.y = element_blank()
    , axis.line.y = element_line(
      colour = "black"
    )
    , axis.text.y = element_text(
      face = "bold.italic"
      , family = "sans"
      , size = 14
      , colour = "black"
    )
    , axis.text.x.top = element_text(
      family = "sans"
      , size = 12
      , margin = margin(
        b = 10
      )
    )
    , axis.title.x = element_blank()
    , legend.position = c(0.20, 0.98)
    , legend.title = element_blank()
    , plot.title = element_text(
      face = "bold"
      , family = "sans"
      , size = 20
      , colour = "black"
    )
    , plot.subtitle = element_text(
      face = "bold"
      , family = "sans"
      , size = 16
      , colour = "grey50"
    )
  ) +
  guides(
    alpha = "none"
    , colour = guide_legend(
      override.aes = list(
        size = 7
      )
      , nrow = 1))+
  theme(panel.grid.major.y =element_line(colour="black"),
        plot.margin=margin(10,50,10,10)
  )


tiff("Figure5_new.tiff", units="in", width=19, height=16, res=300)

# Arrange with widths (2:1 ratio) and p2 at half height
combined_plot <- free(f5_1)/f5_2/f5_3  + 
  plot_layout(heights = c(1.5, 1, 1))
combined_plot

dev.off()


################################ THERE MAY BE DUPLICATION BUT RESTARTING FROM HERE ##############################

############## CREATING PLOTS FROM OUTPUTS _ 2 ##############

rm(list=ls())

#### loading packages
library(tidyr)
library(data.table)
library(dplyr)
library(stringr)
library(ggplot2)

#### !!! note throughout did last minute changes to change e.g. AFRO labelling to AFR
### ideally this would be dealt with at start of analysis


### formatting ###
options(scipen=10000) ## turn off scientific notation
`%!in%` = Negate(`%in%`) ## create function for negating %in%

## getting the different 1 scenario per bug
load("outputs/hospitalC_global_averted.RData")

unique(hospital_c$vaccine_id)

######### cleaning functions ################

keep_vac <- c(               
  "Haemophilus influenzae type B_both_0.9_5 years_All_6, 10, 14 weeks"                                                           
  ,"Salmonella Typhi_0.85_0.7_15 years_All_9 months"                                                                              
  , "Staphylococcus aureus_0.6_0.7_5 years_All_All age groups"                                                               
  , "E. coli - non-diarrhogenic_0.7_0.7_5 years_All except Diarrhoea_All age groups"                                      
  , "ETEC_0.6_0.7_5 years_Diarrhoea_6 months"                                                                   
  , "Group A streptococcus_0.7_0.7_5 years_All_6 weeks"                                                                            
  , "Klebsiella pneumoniae - all_0.7_0.7_5 years_All_All age groups"                                                      
  , "Streptococcus pneumoniae - Improved_both_0.9_5 years_BSI, CNS infections, Cardiac infections, LRI_6 weeks & elderly age group"
  , "Pseudomonas aeruginosa_0.7_0.7_5 years_BSI, LRI and thorax infections_All age groups"                                         
  ,"Enterococcus faecium_0.7_0.7_5 years_All_All age groups"                                                                      
  , "Salmonella paratyphi_0.7_0.7_5 years_All_9 months"                                                                            
  , "Acinetobacter baumannii - all_0.7_0.7_5 years_All_All age groups"                                                            
  , "Mycobacterium tuberculosis - Improved_0.8_0.7_10 years_All_0 weeks + boost every 10 years"                                    
  , "Shigella_0.6_0.7_5 years_All_6 months"
  ,"Non-typhoidal Salmonella_0.8_0.7_5 years_All_6 weeks & 9 months" ) 
### need to specify different ones for productivity as didn't group
## vaccine scenarios like had done for 

## selecting ones that have the theoretical biggest impact (e.g. most age groups/syndromes)

keep_vac_prod <- c(               
  "Haemophilus influenzae type B_0.93_0.9_5 years_All_6, 10, 14 weeks"  
  ,"Haemophilus influenzae type B_0.69_0.9_5 years_All_6, 10, 14 weeks" 
  ,"Streptococcus pneumoniae - Improved_0.7_0.9_5 years_BSI, CNS infections, Cardiac infections, LRI_6 weeks & elderly age group"
  ,"Streptococcus pneumoniae - Improved_0.5_0.9_5 years_BSI, CNS infections, Cardiac infections, LRI_6 weeks & elderly age group"
  ,"Salmonella Typhi_0.85_0.7_15 years_All_9 months"                                                                              
  , "Staphylococcus aureus_0.6_0.7_5 years_All_All age groups"                                                               
  , "E. coli - non-diarrhogenic_0.7_0.7_5 years_All except Diarrhoea_All age groups"                                      
  , "ETEC_0.6_0.7_5 years_Diarrhoea_6 months"                                                                   
  , "Group A streptococcus_0.7_0.7_5 years_All_6 weeks"                                                                            
  , "Klebsiella pneumoniae - all_0.7_0.7_5 years_All_All age groups"                                                      
  , "Pseudomonas aeruginosa_0.7_0.7_5 years_BSI, LRI and thorax infections_All age groups"                                         
  ,"Enterococcus faecium_0.7_0.7_5 years_All_All age groups"                                                                      
  , "Salmonella paratyphi_0.7_0.7_5 years_All_9 months"                                                                            
  , "Acinetobacter baumannii - all_0.7_0.7_5 years_All_All age groups"                                                            
  , "Mycobacterium tuberculosis - Improved_0.8_0.7_10 years_All_0 weeks + boost every 10 years"                                    
  , "Shigella_0.6_0.7_5 years_All_6 months"
  ,"Non-typhoidal Salmonella_0.8_0.7_5 years_All_6 weeks & 9 months" ) 


cleaning.dt <- function(hospital_c){
  global_hospital <- hospital_c[hospital_c$vaccine_id %in% keep_vac]
  
  ### need to specify for E.coli target syndromes so no double counting
  global_hospital[ , flag := 0]
  global_hospital[(vaccine_id=="ETEC_0.6_0.7_5 years_Diarrhoea_6 months" & 
                     Infectious.syndrome!="Diarrhoea")|(
                       vaccine_id=="E. coli - non-diarrhogenic_0.7_0.7_5 years_All except Diarrhoea_All age groups"& 
                         Infectious.syndrome=="Diarrhoea"),
                  flag := 1]
  global_hospital <-  global_hospital[flag==0]
  global_hospital <-  global_hospital[ ,-c("flag")]
  
  
  return(global_hospital)
}

cleaning.dt.prod <- function(hospital_c){
  global_hospital <- hospital_c[hospital_c$vaccine_id %in% keep_vac_prod]
  
  ### need to specify for E.coli target syndromes so no double counting
  global_hospital[ , flag := 0]
  global_hospital[(vaccine_id=="ETEC_0.6_0.7_5 years_Diarrhoea_6 months" & 
                     Infectious.syndrome!="Diarrhoea")|(
                       vaccine_id=="E. coli - non-diarrhogenic_0.7_0.7_5 years_All except Diarrhoea_All age groups"& 
                         Infectious.syndrome=="Diarrhoea"),
                  flag := 1]
  global_hospital <-  global_hospital[flag==0]
  global_hospital <-  global_hospital[ ,-c("flag")]
  
  global_hospital[Infectious.syndrome=="BSI", Infectious.syndrome := "Bloodstream infections"]
  global_hospital[Infectious.syndrome=="UTI", Infectious.syndrome := "Urinary tract infections"]
  global_hospital[Infectious.syndrome=="TB", Infectious.syndrome := "Tuberculosis"]
  global_hospital[Antibiotic.class=="Methicillin", Antibiotic.class := "Penicillin"]
  global_hospital[Antibiotic.class=="Multi-drug resistance excluding extensive drug resistance in TB",
                  Antibiotic.class := "MDR"]
  global_hospital[Antibiotic.class=="Multi-drug resistance in Salmonella Typhi and Paratyphi",
                  Antibiotic.class := "MDR"]
  
  ### change vaccine ids for those we want to combine
  global_hospital[vaccine_id=="Haemophilus influenzae type B_0.93_0.9_5 years_All_6, 10, 14 weeks"|
                    vaccine_id=="Haemophilus influenzae type B_0.69_0.9_5 years_All_6, 10, 14 weeks" ,
                  vaccine_id := "Haemophilus influenzae type B_both_0.9_5 years_All_6, 10, 14 weeks" ]
  
  global_hospital[vaccine_id=="Streptococcus pneumoniae - Improved_0.7_0.9_5 years_BSI, CNS infections, Cardiac infections, LRI_6 weeks & elderly age group"|
                    vaccine_id=="Streptococcus pneumoniae - Improved_0.5_0.9_5 years_BSI, CNS infections, Cardiac infections, LRI_6 weeks & elderly age group",
                  vaccine_id := "Streptococcus pneumoniae - Improved_both_0.9_5 years_BSI, CNS infections, Cardiac infections, LRI_6 weeks & elderly age group"]
  
  return(global_hospital)
}

getting_numbers_bk <- function(regional_hospital_totals){
  ### remove commas from numeric strings
  regional_hospital_totals[ , median_iqr_total_cost := gsub(",", "", gsub("([a-zA-Z]),", "\\1 ", median_iqr_total_cost)) ]
  regional_hospital_totals[ , median_iqr_averted_cost := gsub(",", "", gsub("([a-zA-Z]),", "\\1 ", median_iqr_averted_cost)) ]
  regional_hospital_totals[ , median_iqr_total_days := gsub(",", "", gsub("([a-zA-Z]),", "\\1 ", median_iqr_total_days)) ]
  regional_hospital_totals[ , median_iqr_averted_days := gsub(",", "", gsub("([a-zA-Z]),", "\\1 ", median_iqr_averted_days)) ]
  
  ### separate out numbers in/across brackets
  Separate <- function(...) separate(..., sep = "[^[:alnum:].]+", convert = TRUE)
  
  regional_hospital_totals <- regional_hospital_totals %>%
    Separate(median_iqr_total_cost, into = c("cost_med", "cost_lo", "cost_hi", NA)) %>%
    Separate(median_iqr_averted_cost, into = c("avert_cost_med", "avert_cost_lo", "avert_cost_hi", NA)) %>%
    Separate(median_iqr_total_days, into = c("days_med", "days_lo", "days_hi", NA)) %>%
    Separate(median_iqr_averted_days, into = c("avert_days_med", "avert_days_lo", "avert_days_hi", NA)) %>%
    as.data.table()
  return(regional_hospital_totals)
}

###### ******global totals ********#####################################

#### HOSPITAL COSTS ##########################
load("outputs/hospitalC_global_averted.RData")

hospital_c <- cleaning.dt(hospital_c)

### getting vaccine ids 
vacs <- data.table(vacsid = unique(hospital_c$vaccine_id))
vacs[, c("vaccine_target_disease","efficacy" ,"coverage"  ,
         "duration", "target disease","target population") := tstrsplit(vacsid, "_", fixed=TRUE)]
write.csv(vacs,file="outputs/vaccine_ids.csv")

hospital_c[class=="Fluoroquinolones or MDR in Salmonella",class :="fluoroquinolones & mdr"]
hospital_c[Infectious.syndrome=="BSI", Infectious.syndrome := "Bloodstream infections"]
hospital_c[Infectious.syndrome=="UTI", Infectious.syndrome := "Urinary tract infections"]
hospital_c[Infectious.syndrome=="TB", Infectious.syndrome := "Tuberculosis"]

hospital_c[, c("cost_med","other1") := tstrsplit(median_iqr_total_cost, "(", fixed=TRUE)]
hospital_c[, c("avert_cost_med","other1") := tstrsplit(median_iqr_averted_cost, "(", fixed=TRUE)]

hospital_c[ , cost_med := as.numeric(gsub(",", "", cost_med))]
hospital_c[ , avert_cost_med := as.numeric(gsub(",", "", avert_cost_med))]


classrank <- (hospital_c[, lapply(.SD, sum, na.rm=TRUE),
                         by="class",
                         .SDcols=c("cost_med",
                                   "avert_cost_med")])

pathrank1 <- (hospital_c[, lapply(.SD, sum, na.rm=TRUE),
                         by="Pathogen",
                         .SDcols=c("cost_med",
                                   "avert_cost_med")])


syndrank <- (hospital_c[, lapply(.SD, sum, na.rm=TRUE),
                        by="Infectious.syndrome",
                        .SDcols=c("cost_med",
                                  "avert_cost_med")])


hospital_c <- setorder(hospital_c, cost_med)

## get pathogen in order of cost for plots 
order.test <- hospital_c[, lapply(.SD, sum, na.rm=TRUE),
                         by = c("Pathogen"),
                         .SDcols=c("cost_med")]
order.test <- order.test[order(order.test$cost_med,decreasing = TRUE)] 

hospital_c$Pathogen <- factor(hospital_c$Pathogen , 
                              levels = order.test$Pathogen)

path <- hospital_c$Pathogen

hospital_c$class <- factor(hospital_c$class, levels=unique(hospital_c$class))
class <- hospital_c$class

abxcolours <-
  setNames( c("#660099", "#E69F00", "#56B4E9", "#009E73",
              "#F0E442", "#0072B2", "#D55E00", "#CC79A7")
            , levels(hospital_c$class)  )

ggplot(hospital_c, aes(Pathogen, cost_med,
                       fill=class)) +
  geom_bar(stat="identity")+
  scale_fill_manual(values = abxcolours)+
  theme(axis.text.x = element_text(angle = -90),
        text= element_text(size = 15))+
  scale_y_continuous(label= scales::comma)+
  xlab("Pathogen")+
  ylab("Total Hospital Costs (USD)")+
  labs(fill="Antibiotic Class") 

ggplot(hospital_c, aes(Pathogen, cost_med,
                       fill=Infectious.syndrome)) +
  geom_bar(stat="identity")+
  theme(axis.text.x = element_text(angle = -90),
        text= element_text(size = 15))+
  scale_y_continuous(label= scales::comma)+
  xlab("Pathogen")+
  ylab("Total Hospital Costs (USD)")+
  labs(fill="Syndrome")


############## getting a global median and IQR
load("outputs/total_global_cost.RData")

global_hospital_totals <- cleaning.dt(hospital_c)

### note this is slightly different - median of sums to sum of medians calculated
### above 
global_hospital_totals <- global_hospital_totals %>% 
  group_by(run) %>%
  summarise(cost = sum(hospital_cost, na.rm=TRUE),
            days = sum(hospital_days, na.rm=TRUE),
            cost_avert = sum(avert_hospital_cost,na.rm=TRUE),
            days_avert = sum(avert_hospital_days,na.rm=TRUE)) %>%
  summarise(avert_cost_med = median(cost_avert, na.rm=TRUE),
            avert_cost_lo = quantile(cost_avert,0.25, na.rm = TRUE),
            avert_cost_hi = quantile(cost_avert,0.75, na.rm = TRUE),
            cost_med = median(cost, na.rm=TRUE),
            cost_lo = quantile(cost,0.25, na.rm = TRUE),
            cost_hi = quantile(cost,0.75, na.rm = TRUE),
            avert_days_med = median(days_avert, na.rm=TRUE),
            avert_days_lo = quantile(days_avert,0.25, na.rm = TRUE),
            avert_days_hi = quantile(days_avert,0.75, na.rm = TRUE),
            days_med = median(days, na.rm=TRUE),
            days_lo = quantile(days,0.25, na.rm = TRUE),
            days_hi = quantile(days,0.75, na.rm = TRUE))%>%
  as.data.table()

global_hospital_totals <- global_hospital_totals/1e9

print(global_hospital_totals)

### this is used for top line narrative results in paper
write.csv(global_hospital_totals, file="outputs/global_hospital_totals.csv")

#### Global productivity ##########################

prod_deaths <- read.csv("outputs/productivity_loss_deaths_all.csv")
prod_deaths <- as.data.table(prod_deaths)
### getting one vaccine/one DRI per group
prod_deaths <- cleaning.dt.prod(prod_deaths)

(prod_deaths[, lapply(.SD, sum, na.rm=TRUE),
             .SDcols=c("HC_cost",
                       "averted_HC")])*1/1e9

## setting class the same factor so graphs match


global_prod_deaths <- prod_deaths[, lapply(.SD, sum, na.rm=TRUE),
                                  by = c("Pathogen",
                                         "Antibiotic.class",
                                         "Infectious.syndrome"),
                                  .SDcols=c("HC_cost",
                                            "averted_HC")]

pathrankprod <- (prod_deaths[, lapply(.SD, sum, na.rm=TRUE),
                             by="Pathogen",
                             .SDcols=c("HC_cost","averted_HC")])
### get it so the productivity and hc costs are the same order of pathogens 
### and the same colours for antibiotic classes

abx_dic <- unique(global_prod_deaths[,c("Antibiotic.class")])
abx_dic[Antibiotic.class=="Carbapenems", class:="carbapenems"]
abx_dic[Antibiotic.class=="Fluoroquinolones", class:="fluoroquinolones"]
abx_dic[Antibiotic.class=="Third-generation cephalosporins", class:="3g cephalosporins"]
abx_dic[Antibiotic.class=="Macrolide", class:="macrolides"]
abx_dic[Antibiotic.class=="MDR", class:="mdr"]
abx_dic[Antibiotic.class=="Penicillin", class:="penicillins"]
abx_dic[Antibiotic.class=="Vancomycin", class:="glycopeptides"]

global_prod_deaths <- merge(global_prod_deaths,abx_dic,by="Antibiotic.class")

global_prod_deaths$Pathogen <- factor(global_prod_deaths$Pathogen , 
                                      levels = levels(path))

global_prod_deaths$class <- factor(global_prod_deaths$class , 
                                   levels = levels(class))

ggplot(global_prod_deaths, aes(Pathogen, HC_cost,
                               fill=class)) +
  geom_bar(stat="identity")+
  theme(axis.text.x = element_text(angle = -90),
        text= element_text(size = 15))+
  scale_fill_manual(values = abxcolours)+
  scale_y_continuous(label= scales::comma)+
  xlab("Pathogen")+
  ylab("Total Productivity Losses (USD)")+
  labs(fill='Antibiotic Class') 

ggplot(global_prod_deaths, aes(Pathogen, HC_cost,
                               fill=Infectious.syndrome)) +
  geom_bar(stat="identity")+
  theme(axis.text.x = element_text(angle = -90),
        text= element_text(size = 15))+
  scale_y_continuous(label= scales::comma)+
  xlab("Pathogen")+
  ylab("Total Productivity Losses (USD)")+
  labs(fill='Syndrome')

############******* REGIONAL PLOTS******* ################

########## HOSPITAL ########################
load("outputs/hospitalC_region_averted.RData")

regional_hospital_totals <- cleaning.dt(hospital_c)

regional_hospital_totals[WHO.Region=="PAHO", WHO.Region := "AMRO"]

regional_hospital_totals <- getting_numbers_bk(regional_hospital_totals)
## regional
totals.temp <- (regional_hospital_totals[, lapply(.SD, sum, na.rm=TRUE),
                                         by="WHO.Region",
                                         .SDcols=c("cost_med",
                                                   "avert_cost_med")])

write.csv(totals.temp,file="outputs/regional_median_HC.csv")
rm(totals.temp)

##################### by pathogen - averted hospital costs ####
load("outputs/hospitalC_region_averted_pathogen.RData")

regional_hospital_totals <-    hospital_c[hospital_c$vaccine_id %in% keep_vac]
####!!! don't use for totals pre-vaccine as will have double counting for e.coli
### trying to account for this
### accounting for duplication of E. coli total amounts for ETEC & EXPEC
regional_hospital_totals[vaccine_id=="ETEC_0.6_0.7_5 years_Diarrhoea_6 months",
                         cost_med :=0]
regional_hospital_totals[vaccine_id=="ETEC_0.6_0.7_5 years_Diarrhoea_6 months",
                         cost_lo :=0]
regional_hospital_totals[vaccine_id=="ETEC_0.6_0.7_5 years_Diarrhoea_6 months",
                         cost_hi :=0]
regional_hospital_totals[vaccine_id=="ETEC_0.6_0.7_5 years_Diarrhoea_6 months",
                         days_med :=0]
regional_hospital_totals[vaccine_id=="ETEC_0.6_0.7_5 years_Diarrhoea_6 months",
                         days_lo :=0]
regional_hospital_totals[vaccine_id=="ETEC_0.6_0.7_5 years_Diarrhoea_6 months",
                         days_hi :=0]

regional_hospital_totals <- getting_numbers_bk(regional_hospital_totals)

regional_hospital_totals[WHO.Region=="PAHO", WHO.Region := "AMRO"]

regional_hospital_totals <-    regional_hospital_totals[regional_hospital_totals$vaccine_id %in% keep_vac]
####!!! don't use for totals pre-vaccine as will have double counting for e.coli
### trying to account for this
### accounting for duplication of E. coli total amounts for ETEC & EXPEC
regional_hospital_totals[vaccine_id=="ETEC_0.6_0.7_5 years_Diarrhoea_6 months",
                         cost_med :=0]
regional_hospital_totals[vaccine_id=="ETEC_0.6_0.7_5 years_Diarrhoea_6 months",
                         cost_lo :=0]
regional_hospital_totals[vaccine_id=="ETEC_0.6_0.7_5 years_Diarrhoea_6 months",
                         cost_hi :=0]
regional_hospital_totals[vaccine_id=="ETEC_0.6_0.7_5 years_Diarrhoea_6 months",
                         days_med :=0]
regional_hospital_totals[vaccine_id=="ETEC_0.6_0.7_5 years_Diarrhoea_6 months",
                         days_lo :=0]
regional_hospital_totals[vaccine_id=="ETEC_0.6_0.7_5 years_Diarrhoea_6 months",
                         days_hi :=0]

# save(regional_hospital_totals, file="outputs/regional_hospital_4additionalplots.RData")

total_averted <- regional_hospital_totals[, lapply(.SD, sum, na.rm=TRUE),
                                          by = c("Pathogen"),
                                          .SDcols=c("avert_cost_med")]

total_averted_los <-  regional_hospital_totals[, lapply(.SD, sum, na.rm=TRUE),
                                               by = c("Pathogen"),
                                               .SDcols=c("avert_days_med")]

total_averted <- merge(total_averted, total_averted_los, by="Pathogen")

ggplot(regional_hospital_totals, aes(x = Pathogen, y = avert_cost_med,
                                     fill=WHO.Region)) +
  geom_col(position = "dodge") +
  geom_errorbar(aes(ymin = avert_cost_lo, ymax = avert_cost_hi),
                position = position_dodge(0.9), width = .2)+
  xlab("Resistance Exposure") +
  ylab("Averted Hospital Costs (2019 USD)") +
  theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1),
        text = element_text(size = 15))+
  scale_y_continuous(label= scales::comma)+
  labs(fill="WHO Region")


ggplot(regional_hospital_totals, aes(x = Pathogen, y = avert_days_med,
                                     fill=WHO.Region)) +
  geom_col(position = "dodge") +
  geom_errorbar(aes(ymin = avert_days_lo, ymax = avert_days_hi),
                position = position_dodge(0.9), width = .2)+
  xlab("Pathogen") +
  ylab("Averted Hospital Bed Days") +
  theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1),
        text = element_text(size = 15))+
  scale_y_continuous(label= scales::comma)+
  labs(fill="WHO Region")


#### regional productivity loss #######

prod_deaths <- read.csv("outputs/productivity_loss_deaths_all.csv")
prod_deaths <- as.data.table(prod_deaths)
### getting one vaccine/one DRI per group

prod_deaths <- cleaning.dt.prod(prod_deaths)


# save(prod_deaths, file="outputs/prod_4additionalplots.RData")


### total productivity loss by syndrome
syndromerank <- prod_deaths[, lapply(.SD, sum, na.rm=TRUE),
                            by = c("Infectious.syndrome"),
                            .SDcols=c("HC_cost")]

write.csv(syndromerank, file="outputs/global_syndrome_rankings_productivity.csv")

region_prod  <- prod_deaths[, lapply(.SD, sum, na.rm=TRUE),
                            by = c(
                              ".id"),
                            .SDcols=c("averted_HC")]

syndrome_prod  <-prod_deaths[, lapply(.SD, sum, na.rm=TRUE),
                             by = c(
                               "Infectious.syndrome"),
                             .SDcols=c("averted_HC")]

### averted by pathogen
regional_prod_path  <-prod_deaths[, lapply(.SD, sum, na.rm=TRUE),
                                  by = c("Pathogen",
                                         ".id"),
                                  .SDcols=c("averted_HC")]

ggplot(regional_prod_path, aes(x = Pathogen, y = averted_HC,
                               fill=.id)) +
  geom_col(position = "dodge") +
  xlab("Pathogen") +
  ylab("Averted Productivity Losses (USD)") +
  theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))+
  scale_y_continuous(label= scales::comma)+
  scale_fill_discrete(name = "WHO Region")

####
load("outputs/working_life_years_lost.Rdata")

output_hc <- as.data.table(output_hc)
### getting one vaccine/one DRI per group
output_hc <- cleaning.dt.prod(output_hc)

regional_prod_path  <-output_hc[, lapply(.SD, sum, na.rm=TRUE),
                                by = c("Pathogen",
                                       ".id"),
                                .SDcols=c("averted_HC")]


regional_prod_path[.id=="PAHO",.id := "AMRO"]
regional_prod_path[.id=="AFRO" , .id := "AFR"]
regional_prod_path[.id=="AMRO" , .id := "AMR"]
regional_prod_path[.id=="EMRO" , .id := "EMR"]
regional_prod_path[.id=="EURO" , .id := "EUR"]
regional_prod_path[.id=="SEARO" , .id := "SEAR"]
regional_prod_path[.id=="WPRO" , .id := "WPR"]

ggplot(regional_prod_path  , aes(x = Pathogen, y = averted_HC,
                                 fill=.id)) +
  geom_col(position = "dodge") +
  xlab("Pathogen") +
  ylab("Averted Productivity Losses (WLYL)") +
  theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))+
  scale_y_continuous(label= scales::comma)+
  scale_fill_discrete(name = "WHO Region")


###### table of combining average values per vaccine ####

load("outputs/hospitalC_global_averted.RData")

global_hospital_totals <- cleaning.dt(hospital_c)

global_hospital_totals[class=="Fluoroquinolones or MDR in Salmonella",Antibiotic.class :="any"]
global_hospital_totals[Infectious.syndrome=="BSI", Infectious.syndrome := "Bloodstream infections"]
global_hospital_totals[Infectious.syndrome=="UTI", Infectious.syndrome := "Urinary tract infections"]
global_hospital_totals[Antibiotic.class=="Methicillin", Antibiotic.class := "Penicillin"]
global_hospital_totals[Infectious.syndrome=="TB", Infectious.syndrome := "Tuberculosis"]

global_hospital_totals <- getting_numbers_bk(global_hospital_totals)

vaccine_hospital <- global_hospital_totals[, lapply(.SD, sum, na.rm=TRUE),
                                           by = c("vaccine_id"),
                                           .SDcols=c("avert_cost_med")]

prod_deaths <- read.csv("outputs/productivity_loss_deaths_all.csv")
prod_deaths <- as.data.table(prod_deaths)
### getting one vaccine/one DRI per group
prod_deaths <- cleaning.dt.prod(prod_deaths)

prod_deaths[Infectious.syndrome=="BSI", Infectious.syndrome := "Bloodstream infections"]
prod_deaths[Infectious.syndrome=="UTI", Infectious.syndrome := "Urinary tract infections"]
prod_deaths[Infectious.syndrome=="TB", Infectious.syndrome := "Tuberculosis"]
prod_deaths[Antibiotic.class=="Methicillin", Antibiotic.class := "Penicillin"]
prod_deaths[Antibiotic.class=="Multi-drug resistance excluding extensive drug resistance in TB",
            Antibiotic.class := "MDR"]
prod_deaths[Antibiotic.class=="Multi-drug resistance in Salmonella Typhi and Paratyphi",
            Antibiotic.class := "MDR"]

vaccine_productivity <- prod_deaths[, lapply(.SD, sum, na.rm=TRUE),
                                    by = c("vaccine_id"),
                                    .SDcols=c("averted_HC")]

totalcostingtemp <- merge(vaccine_hospital,vaccine_productivity, by="vaccine_id")

totalcostingtemp[ ,total.cost := avert_cost_med+averted_HC]

write.csv(totalcostingtemp, file="outputs/totalcosting.csv")

#### regional comparing regional friction cost to human capital #########

fc.prod <- read.csv("outputs/vaccine_avertable_FC.csv")

fc.prod <- as.data.table(fc.prod)
### getting one vaccine/one DRI per group
fc.prod <- cleaning.dt.prod(fc.prod)

fc.prod[Infectious.syndrome=="BSI", Infectious.syndrome := "Bloodstream infections"]
fc.prod[Infectious.syndrome=="UTI", Infectious.syndrome := "Urinary tract infections"]
fc.prod[Infectious.syndrome=="TB", Infectious.syndrome := "Tuberculosis"]
fc.prod[Antibiotic.class=="Methicillin", Antibiotic.class := "Penicillin"]
fc.prod[Antibiotic.class=="Multi-drug resistance excluding extensive drug resistance in TB",
        Antibiotic.class := "MDR"]
fc.prod[Antibiotic.class=="Multi-drug resistance in Salmonella Typhi and Paratyphi",
        Antibiotic.class := "MDR"]

(fc.prod[, lapply(.SD, sum, na.rm=TRUE),
         .SDcols=c("wage_loss_total",
                   "wage_loss_averted")])*1/1e9

##### sc2
sc2.prod <- read.csv("outputs/productivity_loss_deaths_all_sc2.csv")

sc2.prod <- as.data.table(sc2.prod)
sc2.prod <- cleaning.dt.prod(sc2.prod)


sc2.prod[Infectious.syndrome=="BSI", Infectious.syndrome := "Bloodstream infections"]
sc2.prod[Infectious.syndrome=="UTI", Infectious.syndrome := "Urinary tract infections"]
sc2.prod[Infectious.syndrome=="TB", Infectious.syndrome := "Tuberculosis"]
sc2.prod[Antibiotic.class=="Methicillin", Antibiotic.class := "Penicillin"]
sc2.prod[Antibiotic.class=="Multi-drug resistance excluding extensive drug resistance in TB",
         Antibiotic.class := "MDR"]
sc2.prod[Antibiotic.class=="Multi-drug resistance in Salmonella Typhi and Paratyphi",
         Antibiotic.class := "MDR"]

(sc2.prod[, lapply(.SD, sum, na.rm=TRUE),
          .SDcols=c("HC_cost",
                    "averted_HC")])*1/1e9

load("outputs/prod_4additionalplots.RData")
sc1.prod <- prod_deaths 

## rename columns pre-merge
names(sc2.prod)[names(sc2.prod) == 'HC_cost'] <- 'sc2_total'
names(sc2.prod)[names(sc2.prod) == 'averted_HC'] <- 'sc2_averted'
names(sc1.prod)[names(sc1.prod) == 'HC_cost'] <- 'sc1_total'
names(sc1.prod)[names(sc1.prod) == 'averted_HC'] <- 'sc1_averted'
names(fc.prod)[names(fc.prod) == 'wage_loss_total'] <- 'fc_total'
names(fc.prod)[names(fc.prod) == 'wage_loss_averted'] <- 'fc_averted'

all.prod.methods <- merge(sc1.prod,sc2.prod, by=c(".id",
                                                  "Pathogen",
                                                  "vaccine_id",
                                                  "Antibiotic.class",
                                                  "Infectious.syndrome"))

all.prod.methods <- merge(all.prod.methods,fc.prod, by.x=c(".id",
                                                           "Pathogen",
                                                           "vaccine_id",
                                                           "Antibiotic.class",
                                                           "Infectious.syndrome"),
                          by.y=c("who.region",
                                 "Pathogen",
                                 "vaccine_id",
                                 "Antibiotic.class",
                                 "Infectious.syndrome"))

# save(all.prod.methods, file="outputs/prod_additional4plots_sc2.RData")

all.prod.methods.long <- data.table::melt(all.prod.methods, id.vars = c(".id",
                                                                        "Pathogen" ,
                                                                        "Infectious.syndrome" ,
                                                                        "vaccine_id"), 
                                          measure.vars = list(c("sc1_averted",
                                                                "sc2_averted",
                                                                "fc_averted"),
                                                              c('sc1_total',
                                                                "sc2_total",
                                                                "fc_total")),
                                          value.name = c('averted','total'),
                                          variable.name="scenario")
all.prod.methods.long[scenario==1 , scenario := "Human Capital Scenario 1"]
all.prod.methods.long[scenario==2 , scenario := "Human Capital Scenario 2"]
all.prod.methods.long[scenario==3 , scenario := "Friction Cost"]


### group by region

all.prod.methods.global <- all.prod.methods[, lapply(.SD, sum, na.rm=TRUE),
                                            by = c(".id"),
                                            .SDcols=c("sc1_total",
                                                      "sc2_total",
                                                      "fc_total")]

all.prod.methods.global.temp <- data.table::melt(all.prod.methods.global, id.vars = c(".id"), 
                                                 measure.vars =c('sc1_total',
                                                                 "sc2_total"),
                                                 value.name = c('total'),
                                                 variable.name="scenario")

all.prod.methods.fc <- all.prod.methods.global[ , c(".id","fc_total")]
all.prod.methods.global <- merge(all.prod.methods.global.temp,all.prod.methods.fc,by=".id",
                                 all=TRUE)

all.prod.methods.global[scenario=="sc1_total" , scenario := "Human Capital Scenario 1"]
all.prod.methods.global[scenario=="sc2_total" , scenario := "Human Capital Scenario 2"]
all.prod.methods.global[scenario=="fc_total" , scenario := "Friction Cost"]

#### radar chart 

all.prod.methods.global[, `Region and Method`:=paste(.id, "-", scenario)]

plt <- ggplot(all.prod.methods.global) +
  # Make custom panel grid
  geom_hline(
    aes(yintercept = y), 
    data.frame(y = c(0:3) * 1000),
    color = "lightgrey"
  ) + 
  # Add bars to represent the cumulative track lengths
  # str_wrap(region, 5) wraps the text so each line has at most 5 characters
  # (but it doesn't break long words!)
  geom_col(
    aes(
      x = reorder(str_wrap(`Region and Method`, 5), total),
      y = total,
      fill = scenario
    ),
    position = "dodge2",
    show.legend = TRUE,
    alpha = .9
  ) +
  
  # Add dots to represent the mean gain
  geom_point(
    aes(
      x = reorder(str_wrap(`Region and Method`, 5),fc_total),
      y = fc_total
    ),
    size = 3,
    color = "gray12"
  ) +
  xlab("Region and Scenario") +
  ylab("Total Labour Productivity Loss from Deaths (USD)") +
  scale_y_continuous(label= scales::comma)+
  # Make it circular!
  coord_polar()

plt

###############################################################
#### newer plots #####
#### code for plotting final outputs by Rhys Kingston 2023
### adapted for this repo


if (!require("pacman")) install.packages("pacman")

pacman::p_load(
  "tidyverse"
  , "readr"
  , "here"
  , "janitor"
  , "scales"
  , "rgdal"
  , "sf"
  , "countrycode"
  , "maps"
  , "rgeos"
  , "devtools"
  , "treemapify"
  , "ggbeeswarm"
  , "viridis"
  , "plotly"
)

########## loading data
load("outputs/regional_hospital_4additionalplots.RData") ## adding sc2 for productivity
load("outputs/prod_additional4plots_sc2.RData")

regional_prod_deaths <- all.prod.methods[, lapply(.SD, sum, na.rm=TRUE),
                                         by = c(".id","Pathogen","vaccine_id"),
                                         .SDcols=c("sc1_total","sc2_total",
                                                   "sc1_averted","sc2_averted")]


### merge together
all.dt <- merge(regional_prod_deaths, regional_hospital_totals, by.x=c(".id","Pathogen","vaccine_id"),
                by.y=c("WHO.Region","Pathogen","vaccine_id"))
### !! note that the total costing set to 0 for ETEC is just because all e.coli is already costed (included diarrheoa)
## in the EXPEC - so applicable to both but removed to remove double counting


## sum across vaccines for pathogens
all.dt <- all.dt[, lapply(.SD, sum, na.rm=TRUE),
                 by = c("Pathogen",
                        ".id"),
                 .SDcols=c(    "sc1_total","sc2_total",
                               "sc1_averted","sc2_averted" ,         
                               "avert_cost_med"  ,"avert_cost_lo",
                               "avert_cost_hi" ,
                               "cost_med",  "cost_lo" ,
                               "cost_hi", "avert_days_med"    ,
                               "avert_days_lo"  ,   "avert_days_hi" ,
                               "days_med"  ,   "days_lo"   , 
                               "days_hi"  )]

all.dt.temp <- all.dt[,.(median_cost_sum_G=sum(cost_med)),
                      by=c("Pathogen")]

all.dt <- merge(all.dt, all.dt.temp, by=c("Pathogen"),
                allow.cartesian = TRUE)

all.dt.temp <- all.dt[,.(median_prod_cost_sum_G=sum(sc1_total)),
                      by=c("Pathogen")]

all.dt <- merge(all.dt, all.dt.temp, by=c("Pathogen"),
                allow.cartesian = TRUE)

all.dt[ , global_total := median_prod_cost_sum_G+median_cost_sum_G]

##### Line plot for hospital costs
options(scipen=10000) ## turn off scientific notation

########### total costs ####################

### renaming regions
all.dt[.id=="AFRO", .id:="AFR"]
all.dt[.id=="AMRO", .id:="AMR"]
all.dt[.id=="EMRO", .id:="EMR"]
all.dt[.id=="EURO", .id:="EUR"]
all.dt[.id=="SEARO", .id:="SEAR"]
all.dt[.id=="WPRO", .id:="WPR"]

# 
tiff("Figure4.tiff", units="in", width=17.5, height=7, res=300)

total.cost.plot <- function (all.dt){
  all.dt %>%
    ggplot(
      mapping = aes(
        x = cost_med
        , size = sc1_total
        , y = fct_reorder(Pathogen,global_total)
        , colour = .id
        , alpha = .8
      )
    ) +
    scale_color_brewer(palette = "Dark2") +
    scale_y_discrete(
      expand = c(
        0
        , 1
      )
    ) +
    geom_beeswarm(
      priority = "ascending"
      , cex = 2
    ) +
    scale_size_continuous(
      range = c(
        2
        , 30
      )
    ) +
    scale_x_log10(
      labels = scales::dollar_format(scale = 1e-6, suffix = "mn")
      , position = "top"
    ) +
    # labs(
    #   title = "Total Costs caused by DRIs, by WHO region"
    #   , subtitle = "Log scaled costs in 2019 USD. The y-axis represents hospital costs. Circles are sized by productivity loss due to DRI deaths, and colour indicates WHO region.\n"
    # ) +
    theme_minimal() +
    theme(
      axis.line.y = element_line(
        colour = "black"
        , linewidth = 1
      )
      , axis.text.y = element_text(
        face = "bold.italic"
        , family = "sans"
        , size = 12
        , colour = "black"
      )
      , axis.text.x.top = element_text(
        family = "sans"
        , size = 10.5
        , margin = margin(
          b = 5
        )
      )
      
      , legend.position = c(0.20, 0.98)
      , legend.title = element_blank()
      , plot.title = element_text(
        face = "bold"
        , family = "sans"
        , size = 20
        , colour = "black"
      )
      , plot.subtitle = element_text(
        face = "bold"
        , family = "sans"
        , size = 16
        , colour = "grey50"
      )
    ) +
    guides(
      alpha = "none"
      , size = "none"
      , colour = guide_legend(
        override.aes = list(
          size = 8
        )
        , nrow = 1) )+
    theme(panel.grid.major.y =element_line(colour="black"))+
    xlab("Cost (USD)")+
    ylab("Microbe")
}


total.cost.plot(all.dt)

dev.off()

##################### FIGURE 3 UPDATED 30-04-2025 #############

rm(list = ls())


#### loading packages
library(tidyr)
library(data.table)
library(dplyr)
library(stringr)
library(ggplot2)
library(patchwork) ## added in journal resub phase

#### !!! note throughout did last minute changes to change e.g. AFRO labelling to AFR
### ideally this would be dealt with at start of analysis



### formatting ###
options(scipen=10000) ## turn off scientific notation
`%!in%` = Negate(`%in%`) ## create function for negating %in%



#### unit cost plotting code #####

### multi panel by gram stain
load("outputs/hospitalC_global_los_npop.RData")
### doesn't matter about vaccine scenarios for this


UNITcost_averted_global <- as.data.table(hospital_c)
# UNITcost_averted_global <- UNITcost_averted_global[Infectious.syndrome !="TB"] ## remove TB !!!
## shorten some long exposure group names
UNITcost_averted_global[class=="Fluoroquinolones or MDR in Salmonella" ,
                         class:="fluoroquinolones/MDR" ]

UNITcost_averted_global[ ,Exposure_Group := paste(Pathogen,";",class)]

## link gram stain
bug_dic <- read.csv("data_all/bug_gram.csv")

UNITcost_averted_global2 <- merge(UNITcost_averted_global,bug_dic, 
                                  by.x="Pathogen",
                                  by.y="bacteria",
                                  all.x=TRUE,
                                  all.y=FALSE)

### fill in by hand miss-matches
UNITcost_averted_global2[Pathogen=="Haemophilus influenzae"
                         |Pathogen=="Non-typhoidal Salmonella" 
                         |Pathogen=="Salmonella Paratyphi"  
                         |Pathogen=="Salmonella Typhi"
                         |Pathogen=="Shigella spp.", gram.stain:="gn"]  

UNITcost_averted_global2[gram.stain=="gn", gram.stain := "Gram-negative"]
UNITcost_averted_global2[gram.stain=="gp", gram.stain := "Gram-positive"]


ggplot(UNITcost_averted_global2, aes(x=interaction(Exposure_Group), y=Median_unitcost, 
                                     fill=Infectious.syndrome)) + 
  geom_bar(position=position_dodge(), stat="identity",
           colour="black", # Use black outlines,
           size=.3) +      # Thinner lines
  geom_errorbar(aes(ymin=LOWIQR_unitcost, ymax=HIGIQR_unitcost),
                size=.3,    # Thinner lines
                width=.2,
                position=position_dodge(.9)) +
  xlab("Resistance & Bacterial Exposure") +
  ylab("Hospital Cost per Case (2019 USD)") +
  scale_fill_brewer(palette = "Paired")+
  theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1),
        text = element_text(size = 15))+
  coord_cartesian(ylim=c(NA, 10000), expand = FALSE)+
  facet_wrap(~gram.stain, scales="free")



########## new plot 14/12

## update syndrome labels
UNITcost_averted_global2[Infectious.syndrome=="Bacterial skin infections" ,
                         Infectious.syndrome:="Skin"]
UNITcost_averted_global2[Infectious.syndrome=="Cardiac infections" ,
                         Infectious.syndrome:="Cardiac"]
UNITcost_averted_global2[Infectious.syndrome=="LRI and thorax infections" | 
                           Infectious.syndrome=="TB" ,
                         Infectious.syndrome:="RTI/TB"]
UNITcost_averted_global2[Infectious.syndrome=="Bone and joint infections"   ,
                         Infectious.syndrome:="Bone & Joint"]
UNITcost_averted_global2[Infectious.syndrome=="Intra-abdominal infections" ,
                         Infectious.syndrome:="IAI"]
UNITcost_averted_global2[Infectious.syndrome=="CNS infections"  ,
                         Infectious.syndrome:="CNS"]
UNITcost_averted_global2[Infectious.syndrome=="Typhoid, paratyphoid, and iNTS",
                         Infectious.syndrome:="(para-)/Typhoid/iNTS"]

# ## order
temp <- UNITcost_averted_global2 %>% 
  group_by(Infectious.syndrome) %>% summarise(mM=mean(Median_unitcost))

UNITcost_averted_global2[ ,Infectious.syndrome:=factor(Infectious.syndrome,
                                                       levels=c("RTI/TB",
                                                                "UTI",
                                                                "Cardiac",
                                                                "CNS",
                                                                "Diarrhoea",
                                                                "Skin",
                                                                "Bone & Joint",
                                                                "BSI",
                                                                "IAI",
                                                                "(para-)/Typhoid/iNTS"))]


### addition 2025 - relabelling to make plot look better
UNITcost_averted_global2[ Infectious.syndrome== "(para-)/Typhoid/iNTS",
                          Infectious.syndrome:= "(p-)T & iNTS"]

ggplot(UNITcost_averted_global2, aes(x=interaction(Exposure_Group), y=Median_unitcost,
                                     fill=class)) + 
  geom_bar(position=position_dodge(), stat="identity",
           colour="black", # Use black outlines,
           size=.3) +      # Thinner lines
  geom_errorbar(aes(ymin=LOWIQR_unitcost, ymax=HIGIQR_unitcost),
                size=.3,    # Thinner lines
                width=.2,
                position=position_dodge(.9))+
  coord_flip()+
  facet_grid(~Infectious.syndrome, switch = "y", scales = "free_y") +
  theme(strip.background = element_blank(),
        strip.placement = "inside",
        axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1),
        panel.spacing = unit(0.35, "lines"),
        plot.title = element_text(size = 12),
        axis.title.x = element_text(size = 12),
        axis.title.y = element_text(size = 12),
        legend.position = "none") +
  ylab("Hospital cost per case ($)")+
  xlab("Exposure group (bacteria;ABR)")+
  ggtitle("Syndrome")

TB <- UNITcost_averted_global2[Pathogen=="Mycobacterium tuberculosis"]
nonTB <- UNITcost_averted_global2[Pathogen!="Mycobacterium tuberculosis"]


p1 <- ggplot(nonTB, aes(x=interaction(Exposure_Group), y=Median_unitcost,
                        fill=class)) + 
  geom_bar(position=position_dodge(), stat="identity",
           colour="black", # Use black outlines,
           size=.3) +      # Thinner lines
  geom_errorbar(aes(ymin=LOWIQR_unitcost, ymax=HIGIQR_unitcost),
                size=.3,    # Thinner lines
                width=.2,
                position=position_dodge(.9))+
  coord_flip()+
  facet_grid(~Infectious.syndrome, switch = "y", scales = "free_y") +
  theme(strip.background = element_blank(),
        strip.placement = "inside",
        axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1),
        panel.spacing = unit(0.35, "lines"),
        plot.title = element_text(size = 12),
        axis.title.x = element_text(size = 12),
        axis.title.y = element_text(size = 12),
        legend.position = "none") +
  ylab("Hospital cost per case ($)")+
  xlab("Exposure group (bacteria;ABR)")+
  ggtitle("(A) Syndrome")

### rename TB to help plot saving
TB[Exposure_Group=="Mycobacterium tuberculosis ; mdr", Exposure_Group:="MDR TB"]

p2 <- ggplot(TB, aes(x=interaction(Exposure_Group), y=Median_unitcost,
                     fill=class)) + 
  geom_bar(position=position_dodge(), stat="identity",
           colour="black", # Use black outlines,
           size=.3) +      # Thinner lines
  geom_errorbar(aes(ymin=LOWIQR_unitcost, ymax=HIGIQR_unitcost),
                size=.3,    # Thinner lines
                width=.2,
                position=position_dodge(.9))+
  coord_flip()+
  facet_grid(~Infectious.syndrome, switch = "y", scales = "free_y") +
  theme(strip.background = element_blank(),
        strip.placement = "inside",
        axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1),
        panel.spacing = unit(0.35, "lines"),
        plot.title = element_text(size = 12),
        axis.title.x = element_text(size = 12),
        axis.title.y = element_text(size = 12),
        legend.position = "none") +
  ylab("Hospital cost per case ($)")+
  xlab("Exposure group")+
  ggtitle("(B) Syndrome")

# Wrap p2 inside wrap_elements() to shrink its height
p2_wrapped <- wrap_elements(full = p2) / plot_spacer() # Spacer makes p2 take less space

tiff("Figure3.tiff", units="in", width=17.5, height=7, res=300)

# Arrange with widths (2:1 ratio) and p2 at half height
combined_plot <- (p1 | p2_wrapped) + 
  plot_layout(widths = c(3.5, 1), heights = c(1, 0.5))
combined_plot

dev.off()
# ###############
# 
# ggplot(nonTB, aes(x=interaction(Exposure_Group), y=Median_unitcost,
#                   fill=class)) + 
#   geom_bar(position=position_dodge(), stat="identity",
#            colour="black", # Use black outlines,
#            size=.3) +      # Thinner lines
#   geom_errorbar(aes(ymin=LOWIQR_unitcost, ymax=HIGIQR_unitcost),
#                 size=.3,    # Thinner lines
#                 width=.2,
#                 position=position_dodge(.9))+
#   coord_flip()+
#   facet_grid(~Infectious.syndrome, switch = "x", scales = "free_y") +
#   scale_y_continuous(position = "right")+
#   theme(strip.background = element_blank(),
#         strip.placement = "inside",
#         axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1),
#         panel.spacing = unit(0.35, "lines"),
#         plot.title = element_text(size = 12, vjust=10),
#         axis.title.x = element_text(size = 12),
#         axis.title.y = element_text(size = 12),
#         legend.position = "none") +
#   ylab("Hospital cost per case ($)")+
#   xlab("Exposure group")+
#   ggtitle("Syndrome")
#### unit cost plotting code #####

### multi panel by gram stain
load("outputs/hospitalC_global_los_npop.RData")
### doesn't matter about vaccine scenarios for this


UNITcost_averted_global <- as.data.table(hospital_c)
# UNITcost_averted_global <- UNITcost_averted_global[Infectious.syndrome !="TB"] ## remove TB !!!
## shorten some long exposure group names
UNITcost_averted_global[class=="Fluoroquinolones or MDR in Salmonella" ,
                         class:="fluoroquinolones/MDR" ]

UNITcost_averted_global[ ,Exposure_Group := paste(Pathogen,";",class)]

## link gram stain
bug_dic <- read.csv("data_all/bug_gram.csv")

UNITcost_averted_global2 <- merge(UNITcost_averted_global,bug_dic, 
                                  by.x="Pathogen",
                                  by.y="bacteria",
                                  all.x=TRUE,
                                  all.y=FALSE)

### fill in by hand miss-matches
UNITcost_averted_global2[Pathogen=="Haemophilus influenzae"
                         |Pathogen=="Non-typhoidal Salmonella" 
                         |Pathogen=="Salmonella Paratyphi"  
                         |Pathogen=="Salmonella Typhi"
                         |Pathogen=="Shigella spp.", gram.stain:="gn"]  

UNITcost_averted_global2[gram.stain=="gn", gram.stain := "Gram-negative"]
UNITcost_averted_global2[gram.stain=="gp", gram.stain := "Gram-positive"]


ggplot(UNITcost_averted_global2, aes(x=interaction(Exposure_Group), y=Median_unitcost, 
                                     fill=Infectious.syndrome)) + 
  geom_bar(position=position_dodge(), stat="identity",
           colour="black", # Use black outlines,
           size=.3) +      # Thinner lines
  geom_errorbar(aes(ymin=LOWIQR_unitcost, ymax=HIGIQR_unitcost),
                size=.3,    # Thinner lines
                width=.2,
                position=position_dodge(.9)) +
  xlab("Resistance & Bacterial Exposure") +
  ylab("Hospital Cost per Case (2019 USD)") +
  scale_fill_brewer(palette = "Paired")+
  theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1),
        text = element_text(size = 15))+
  coord_cartesian(ylim=c(NA, 10000), expand = FALSE)+
  facet_wrap(~gram.stain, scales="free")



########## new plot 14/12

## update syndrome labels
UNITcost_averted_global2[Infectious.syndrome=="Bacterial skin infections" ,
                         Infectious.syndrome:="Skin"]
UNITcost_averted_global2[Infectious.syndrome=="Cardiac infections" ,
                         Infectious.syndrome:="Cardiac"]
UNITcost_averted_global2[Infectious.syndrome=="LRI and thorax infections" | 
                           Infectious.syndrome=="TB" ,
                         Infectious.syndrome:="RTI/TB"]
UNITcost_averted_global2[Infectious.syndrome=="Bone and joint infections"   ,
                         Infectious.syndrome:="Bone & Joint"]
UNITcost_averted_global2[Infectious.syndrome=="Intra-abdominal infections" ,
                         Infectious.syndrome:="IAI"]
UNITcost_averted_global2[Infectious.syndrome=="CNS infections"  ,
                         Infectious.syndrome:="CNS"]
UNITcost_averted_global2[Infectious.syndrome=="Typhoid, paratyphoid, and iNTS",
                         Infectious.syndrome:="(para-)/Typhoid/iNTS"]

# ## order
temp <- UNITcost_averted_global2 %>% 
  group_by(Infectious.syndrome) %>% summarise(mM=mean(Median_unitcost))

UNITcost_averted_global2[ ,Infectious.syndrome:=factor(Infectious.syndrome,
                                                       levels=c("RTI/TB",
                                                                "UTI",
                                                                "Cardiac",
                                                                "CNS",
                                                                "Diarrhoea",
                                                                "Skin",
                                                                "Bone & Joint",
                                                                "BSI",
                                                                "IAI",
                                                                "(para-)/Typhoid/iNTS"))]


### addition 2025 - relabelling to make plot look better
UNITcost_averted_global2[ Infectious.syndrome== "(para-)/Typhoid/iNTS",
                          Infectious.syndrome:= "(p-)T & iNTS"]

ggplot(UNITcost_averted_global2, aes(x=interaction(Exposure_Group), y=Median_unitcost,
                                     fill=class)) + 
  geom_bar(position=position_dodge(), stat="identity",
           colour="black", # Use black outlines,
           size=.3) +      # Thinner lines
  geom_errorbar(aes(ymin=LOWIQR_unitcost, ymax=HIGIQR_unitcost),
                size=.3,    # Thinner lines
                width=.2,
                position=position_dodge(.9))+
  coord_flip()+
  facet_grid(~Infectious.syndrome, switch = "y", scales = "free_y") +
  theme(strip.background = element_blank(),
        strip.placement = "inside",
        axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1),
        panel.spacing = unit(0.35, "lines"),
        plot.title = element_text(size = 12),
        axis.title.x = element_text(size = 12),
        axis.title.y = element_text(size = 12),
        legend.position = "none") +
  ylab("Hospital cost per case ($)")+
  xlab("Exposure group (bacteria;ABR)")+
  ggtitle("Syndrome")

TB <- UNITcost_averted_global2[Pathogen=="Mycobacterium tuberculosis"]
nonTB <- UNITcost_averted_global2[Pathogen!="Mycobacterium tuberculosis"]


p1 <- ggplot(nonTB, aes(x=interaction(Exposure_Group), y=Median_unitcost,
                        fill=class)) + 
  geom_bar(position=position_dodge(), stat="identity",
           colour="black", # Use black outlines,
           size=.3) +      # Thinner lines
  geom_errorbar(aes(ymin=LOWIQR_unitcost, ymax=HIGIQR_unitcost),
                size=.3,    # Thinner lines
                width=.2,
                position=position_dodge(.9))+
  coord_flip()+
  facet_grid(~Infectious.syndrome, switch = "y", scales = "free_y") +
  theme(strip.background = element_blank(),
        strip.placement = "inside",
        axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1),
        panel.spacing = unit(0.35, "lines"),
        plot.title = element_text(size = 12),
        axis.title.x = element_text(size = 12),
        axis.title.y = element_text(size = 12),
        legend.position = "none") +
  ylab("Hospital cost per case ($)")+
  xlab("Exposure group (bacteria;ABR)")+
  ggtitle("(A) Syndrome")

### rename TB to help plot saving
TB[Exposure_Group=="Mycobacterium tuberculosis ; mdr", Exposure_Group:="MDR TB"]

p2 <- ggplot(TB, aes(x=interaction(Exposure_Group), y=Median_unitcost,
                     fill=class)) + 
  geom_bar(position=position_dodge(), stat="identity",
           colour="black", # Use black outlines,
           size=.3) +      # Thinner lines
  geom_errorbar(aes(ymin=LOWIQR_unitcost, ymax=HIGIQR_unitcost),
                size=.3,    # Thinner lines
                width=.2,
                position=position_dodge(.9))+
  coord_flip()+
  facet_grid(~Infectious.syndrome, switch = "y", scales = "free_y") +
  theme(strip.background = element_blank(),
        strip.placement = "inside",
        axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1),
        panel.spacing = unit(0.35, "lines"),
        plot.title = element_text(size = 12),
        axis.title.x = element_text(size = 12),
        axis.title.y = element_text(size = 12),
        legend.position = "none") +
  ylab("Hospital cost per case ($)")+
  xlab("Exposure group")+
  ggtitle("(B) Syndrome")

# Wrap p2 inside wrap_elements() to shrink its height
p2_wrapped <- wrap_elements(full = p2) / plot_spacer() # Spacer makes p2 take less space

tiff("Figure3.tiff", units="in", width=17.5, height=7, res=300)

# Arrange with widths (2:1 ratio) and p2 at half height
combined_plot <- (p1 | p2_wrapped) + 
  plot_layout(widths = c(3.5, 1), heights = c(1, 0.5))
combined_plot

dev.off()
# ###############
# 
# ggplot(nonTB, aes(x=interaction(Exposure_Group), y=Median_unitcost,
#                   fill=class)) + 
#   geom_bar(position=position_dodge(), stat="identity",
#            colour="black", # Use black outlines,
#            size=.3) +      # Thinner lines
#   geom_errorbar(aes(ymin=LOWIQR_unitcost, ymax=HIGIQR_unitcost),
#                 size=.3,    # Thinner lines
#                 width=.2,
#                 position=position_dodge(.9))+
#   coord_flip()+
#   facet_grid(~Infectious.syndrome, switch = "x", scales = "free_y") +
#   scale_y_continuous(position = "right")+
#   theme(strip.background = element_blank(),
#         strip.placement = "inside",
#         axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1),
#         panel.spacing = unit(0.35, "lines"),
#         plot.title = element_text(size = 12, vjust=10),
#         axis.title.x = element_text(size = 12),
#         axis.title.y = element_text(size = 12),
#         legend.position = "none") +
#   ylab("Hospital cost per case ($)")+
#   xlab("Exposure group")+
#   ggtitle("Syndrome")

#### unit cost plotting code #####

### multi panel by gram stain
load("outputs/hospitalC_global_los_npop.RData")
### doesn't matter about vaccine scenarios for this


UNITcost_averted_global <- as.data.table(hospital_c)
# UNITcost_averted_global <- UNITcost_averted_global[Infectious.syndrome !="TB"] ## remove TB !!!
## shorten some long exposure group names
UNITcost_averted_global[class=="Fluoroquinolones or MDR in Salmonella" ,
                        class:="fluoroquinolones/MDR" ]

UNITcost_averted_global[ ,Exposure_Group := paste(Pathogen,";",class)]

## link gram stain
bug_dic <- read.csv("data_all/bug_gram.csv")

UNITcost_averted_global2 <- merge(UNITcost_averted_global,bug_dic, 
                                  by.x="Pathogen",
                                  by.y="bacteria",
                                  all.x=TRUE,
                                  all.y=FALSE)

### fill in by hand miss-matches
UNITcost_averted_global2[Pathogen=="Haemophilus influenzae"
                         |Pathogen=="Non-typhoidal Salmonella" 
                         |Pathogen=="Salmonella Paratyphi"  
                         |Pathogen=="Salmonella Typhi"
                         |Pathogen=="Shigella spp.", gram.stain:="gn"]  

UNITcost_averted_global2[gram.stain=="gn", gram.stain := "Gram-negative"]
UNITcost_averted_global2[gram.stain=="gp", gram.stain := "Gram-positive"]


ggplot(UNITcost_averted_global2, aes(x=interaction(Exposure_Group), y=Median_unitcost, 
                                     fill=Infectious.syndrome)) + 
  geom_bar(position=position_dodge(), stat="identity",
           colour="black", # Use black outlines,
           size=.3) +      # Thinner lines
  geom_errorbar(aes(ymin=LOWIQR_unitcost, ymax=HIGIQR_unitcost),
                size=.3,    # Thinner lines
                width=.2,
                position=position_dodge(.9)) +
  xlab("Resistance & Bacterial Exposure") +
  ylab("Hospital Cost per Case (2019 USD)") +
  scale_fill_brewer(palette = "Paired")+
  theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1),
        text = element_text(size = 15))+
  coord_cartesian(ylim=c(NA, 10000), expand = FALSE)+
  facet_wrap(~gram.stain, scales="free")



########## new plot 14/12

## update syndrome labels
UNITcost_averted_global2[Infectious.syndrome=="Bacterial skin infections" ,
                         Infectious.syndrome:="Skin"]
UNITcost_averted_global2[Infectious.syndrome=="Cardiac infections" ,
                         Infectious.syndrome:="Cardiac"]
UNITcost_averted_global2[Infectious.syndrome=="LRI and thorax infections" | 
                           Infectious.syndrome=="TB" ,
                         Infectious.syndrome:="RTI/TB"]
UNITcost_averted_global2[Infectious.syndrome=="Bone and joint infections"   ,
                         Infectious.syndrome:="Bone & Joint"]
UNITcost_averted_global2[Infectious.syndrome=="Intra-abdominal infections" ,
                         Infectious.syndrome:="IAI"]
UNITcost_averted_global2[Infectious.syndrome=="CNS infections"  ,
                         Infectious.syndrome:="CNS"]
UNITcost_averted_global2[Infectious.syndrome=="Typhoid, paratyphoid, and iNTS",
                         Infectious.syndrome:="(para-)/Typhoid/iNTS"]

# ## order
temp <- UNITcost_averted_global2 %>% 
  group_by(Infectious.syndrome) %>% summarise(mM=mean(Median_unitcost))

UNITcost_averted_global2[ ,Infectious.syndrome:=factor(Infectious.syndrome,
                                                       levels=c("RTI/TB",
                                                                "UTI",
                                                                "Cardiac",
                                                                "CNS",
                                                                "Diarrhoea",
                                                                "Skin",
                                                                "Bone & Joint",
                                                                "BSI",
                                                                "IAI",
                                                                "(para-)/Typhoid/iNTS"))]


### addition 2025 - relabelling to make plot look better
UNITcost_averted_global2[ Infectious.syndrome== "(para-)/Typhoid/iNTS",
                          Infectious.syndrome:= "(p-)T & iNTS"]

ggplot(UNITcost_averted_global2, aes(x=interaction(Exposure_Group), y=Median_unitcost,
                                     fill=class)) + 
  geom_bar(position=position_dodge(), stat="identity",
           colour="black", # Use black outlines,
           size=.3) +      # Thinner lines
  geom_errorbar(aes(ymin=LOWIQR_unitcost, ymax=HIGIQR_unitcost),
                size=.3,    # Thinner lines
                width=.2,
                position=position_dodge(.9))+
  coord_flip()+
  facet_grid(~Infectious.syndrome, switch = "y", scales = "free_y") +
  theme(strip.background = element_blank(),
        strip.placement = "inside",
        axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1),
        panel.spacing = unit(0.35, "lines"),
        plot.title = element_text(size = 12),
        axis.title.x = element_text(size = 12),
        axis.title.y = element_text(size = 12),
        legend.position = "none") +
  ylab("Hospital cost per case ($)")+
  xlab("Exposure group (bacteria;ABR)")+
  ggtitle("Syndrome")

TB <- UNITcost_averted_global2[Pathogen=="Mycobacterium tuberculosis"]
nonTB <- UNITcost_averted_global2[Pathogen!="Mycobacterium tuberculosis"]


p1 <- ggplot(nonTB, aes(x=interaction(Exposure_Group), y=Median_unitcost,
                        fill=class)) + 
  geom_bar(position=position_dodge(), stat="identity",
           colour="black", # Use black outlines,
           size=.3) +      # Thinner lines
  geom_errorbar(aes(ymin=LOWIQR_unitcost, ymax=HIGIQR_unitcost),
                size=.3,    # Thinner lines
                width=.2,
                position=position_dodge(.9))+
  coord_flip()+
  facet_grid(~Infectious.syndrome, switch = "y", scales = "free_y") +
  theme(strip.background = element_blank(),
        strip.placement = "inside",
        axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1),
        panel.spacing = unit(0.35, "lines"),
        plot.title = element_text(size = 12),
        axis.title.x = element_text(size = 12),
        axis.title.y = element_text(size = 12),
        legend.position = "none") +
  ylab("Hospital cost per case ($)")+
  xlab("Exposure group (bacteria;ABR)")+
  ggtitle("(A) Syndrome")

### rename TB to help plot saving
TB[Exposure_Group=="Mycobacterium tuberculosis ; mdr", Exposure_Group:="MDR TB"]

p2 <- ggplot(TB, aes(x=interaction(Exposure_Group), y=Median_unitcost,
                     fill=class)) + 
  geom_bar(position=position_dodge(), stat="identity",
           colour="black", # Use black outlines,
           size=.3) +      # Thinner lines
  geom_errorbar(aes(ymin=LOWIQR_unitcost, ymax=HIGIQR_unitcost),
                size=.3,    # Thinner lines
                width=.2,
                position=position_dodge(.9))+
  coord_flip()+
  facet_grid(~Infectious.syndrome, switch = "y", scales = "free_y") +
  theme(strip.background = element_blank(),
        strip.placement = "inside",
        axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1),
        panel.spacing = unit(0.35, "lines"),
        plot.title = element_text(size = 12),
        axis.title.x = element_text(size = 12),
        axis.title.y = element_text(size = 12),
        legend.position = "none") +
  ylab("Hospital cost per case ($)")+
  xlab("Exposure group")+
  ggtitle("(B) Syndrome")

# Wrap p2 inside wrap_elements() to shrink its height
p2_wrapped <- wrap_elements(full = p2) / plot_spacer() # Spacer makes p2 take less space

tiff("Figure3.tiff", units="in", width=17.5, height=7, res=300)

# Arrange with widths (2:1 ratio) and p2 at half height
combined_plot <- (p1 | p2_wrapped) + 
  plot_layout(widths = c(3.5, 1), heights = c(1, 0.5))
combined_plot

dev.off()
# ###############
# 
# ggplot(nonTB, aes(x=interaction(Exposure_Group), y=Median_unitcost,
#                   fill=class)) + 
#   geom_bar(position=position_dodge(), stat="identity",
#            colour="black", # Use black outlines,
#            size=.3) +      # Thinner lines
#   geom_errorbar(aes(ymin=LOWIQR_unitcost, ymax=HIGIQR_unitcost),
#                 size=.3,    # Thinner lines
#                 width=.2,
#                 position=position_dodge(.9))+
#   coord_flip()+
#   facet_grid(~Infectious.syndrome, switch = "x", scales = "free_y") +
#   scale_y_continuous(position = "right")+
#   theme(strip.background = element_blank(),
#         strip.placement = "inside",
#         axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1),
#         panel.spacing = unit(0.35, "lines"),
#         plot.title = element_text(size = 12, vjust=10),
#         axis.title.x = element_text(size = 12),
#         axis.title.y = element_text(size = 12),
#         legend.position = "none") +
#   ylab("Hospital cost per case ($)")+
#   xlab("Exposure group")+
#   ggtitle("Syndrome")
